% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClusterings.R
\docType{methods}
\name{addClusterings}
\alias{addClusterings}
\alias{removeClusterings}
\alias{addClusterings,ClusterExperiment,matrix-method}
\alias{addClusterings,ClusterExperiment,ClusterExperiment-method}
\alias{addClusterings,ClusterExperiment,vector-method}
\title{Add clusterings to ClusterExperiment object}
\usage{
\S4method{addClusterings}{ClusterExperiment,matrix}(x, y,
  clusterTypes = "User", clusterLabels = NULL, clusterLegend = NULL)

\S4method{addClusterings}{ClusterExperiment,ClusterExperiment}(x, y)

\S4method{addClusterings}{ClusterExperiment,vector}(x, y, makePrimary = FALSE,
  ...)
}
\arguments{
\item{makePrimary}{whether to make the added cluster the primary cluster (only relevant if \code{y} is a vector)}

\item{...}{For \code{addClusterings}, passed to signature 
\code{ClusterExperiment,matrix}. For \code{[} (subsetting), passed to 
\code{SingleCellExperiment} subsetting function.}
}
\value{
A \code{ClusterExperiment} object.
}
\description{
Function for adding new clusterings in form of vector (single cluster) or matrix (multiple clusterings) to an existing ClusterExperiment object
}
\examples{
data(simData)

cl1 <- clusterSingle(simData, subsample=FALSE,
sequential=FALSE, mainClusterArgs=list(clusterArgs=list(k=3), clusterFunction="pam"))
cl2 <- clusterSingle(simData, subsample=FALSE,
sequential=FALSE, mainClusterArgs=list(clusterArgs=list(k=3), clusterFunction="pam"))

addClusterings(cl1, cl2)
}
