% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelper.R
\docType{methods}
\name{removeUnclustered,ClusterExperiment-method}
\alias{addClusters}
\alias{addClusters,ClusterExperiment,ClusterExperiment-method}
\alias{addClusters,ClusterExperiment,numeric-method}
\alias{removeClusters}
\alias{removeClusters,ClusterExperiment,character-method}
\alias{removeClusters,ClusterExperiment,numeric-method}
\alias{removeUnclustered,ClusterExperiment-method}
\title{Function to add clusters to an existing ClusterExperiment object}
\usage{
\S4method{removeUnclustered}{ClusterExperiment}(x)

\S4method{removeClusters}{ClusterExperiment,character}(x, whichRemove,
  exactMatch = TRUE)

\S4method{removeClusters}{ClusterExperiment,numeric}(x, whichRemove)

\S4method{addClusters}{ClusterExperiment,ClusterExperiment}(x, y)

\S4method{addClusters}{ClusterExperiment,matrix}(x, y, clusterType = "User")

\S4method{addClusters}{ClusterExperiment,numeric}(x, y, ...)
}
\arguments{
\item{x}{a ClusterExperiment Object}

\item{whichRemove}{which clusters to remove. Can be numeric or character. If
numeric, must give indices of clusterMatrix(x) to remove. If character,
should match a clusterType of x}

\item{exactMatch}{logical. Whether the whichRemove must exactly match a value
of clusterType(x). Only relevant if whichRemove is character.}

\item{y}{additional clusters to add to x. Can be ClusterExperiment Object or
a matrix/vector of clusters}
}
\details{
removeUnclustered removes all samples that are unclustered
(i.e. -1 or -2 assignment) in the primaryCluster of x (so they may be
unclustered in other clusters found in clusterMatrix(x))

removeClusters removes the clusters given by whichRemove. If all
clusters are implied, then returns a SummarizedExperiment Object. If the
primaryCluster is one of the clusters removed, the primaryClusterIndex is set
to 1 and the dendrogram and cooccurance matrix are discarded and orderSamples
is set to 1:NCOL(x).

addClusters adds y to x, and is thus not symmetric in the two
arguments. In particular, the primaryCluster and all of its supporting
information (dendrogram, coClustering, and orderSamples) are all kept from
the x object, even if y is a ClusterExperiment.
}

