% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorDefinitions.R
\docType{data}
\name{bigPalette}
\alias{bigPalette}
\alias{showThisPal}
\title{Large palette of colors}
\format{An object of class \code{character} of length 60.}
\usage{
bigPalette

showThisPal()
}
\description{
Large palette of colors
}
\details{
\code{bigPalette} is a long palette of colors (length 62) used by \code{\link{plotTracking}} and
accompanying functions. \code{showThisPal} creates plot that gives index of each color in bigPalette.
}
\keyword{datasets}

