% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusterAll.R
\name{clusterAll}
\alias{clusterAll}
\title{General wrap-around for all clustering methods we're trying.}
\usage{
clusterAll(x, subsample = TRUE, sequential = FALSE,
  clusterMethod = c("tight", "hierarchical", "pam", "kmeans"),
  clusterArgs = NULL, subsampleArgs = NULL, seqArgs = NULL)
}
\arguments{
\item{x}{the data on which to run the clustering (samples in rows)}

\item{subsample}{logical as to whether to subsample to get the distance matrix; otherwise the distance matrix is dist(x)}

\item{sequential}{logical whether to use the sequential strategy.}

\item{clusterMethod}{method for clustering distance matrix (either from subsampling or from dist(x)); if not subsampling, must be “pam”. Passed to clusterDMat option 'method'.}

\item{clusterArgs}{list of additional arguments to be passed to clusterDMat}

\item{subsampleArgs}{list of arguments to be passed to subsamplingClustering}
}
\value{
A list with values
\itemize{
\item{\code{clustering}}{a vector of length equal to nrows(x) giving the integer-valued cluster ids for each sample. The integer values are assigned in the order that the clusters were found, if sequential=TRUE. "-1" indicates the sample was not clustered.}
\item{\code{clusterInfo}}{if sequential=TRUE and clusters were successfully found, a matrix of information regarding the algorithm behavior for each cluster (the starting and stopping K for each cluster, and the number of iterations for each cluster).}
\item{\code{whyStop}}{if sequential=TRUE and clusters were successfully found, a character string explaining what triggered the algorithm to stop.}
}
}
\description{
Given a nxp data matrix, this function will find clusters
}
\details{
If sequential=TRUE, the sequential clustering controls the 'k' argument of the underlying pam/kmeans clustering so setting 'k=' in the list given to clusterArgs will not do anything and will produce a warning to that effect.
}
\examples{
set.seed(44261)
clustSeqHier_v2<-clusterAll(z,clusterMethod="hierarchical",sequential=TRUE,subsample=TRUE,
	subsampleArgs=list(resamp.n=100,samp.p=0.7),
	subsampleClusterArgs=list(nstart=10),
	seqArgs=list(beta=0.8,k0=10),
	clusterArgs=list(min.size=5))
#use clusterAll to do just clustering k=3 with no subsampling
clustNothing<-clusterAll(z,clusterMethod="pam",subsample=FALSE,sequential=FALSE
	clusterArgs=list(k=3))
}

