% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R, R/AllHelper.R
\docType{class}
\name{ClusterCells-class}
\alias{ClusterCells}
\alias{ClusterCells-class}
\alias{addClusters,ClusterCells,ClusterCells-method}
\alias{addClusters,ClusterCells,matrix-method}
\alias{addClusters,ClusterCells,numeric-method}
\alias{allClusters}
\alias{allClusters,ClusterCells-method}
\alias{clusterCells}
\alias{clusterCells,SummarizedExperiment,character-method}
\alias{clusterCells,SummarizedExperiment,factor-method}
\alias{clusterCells,SummarizedExperiment,matrix-method}
\alias{clusterCells,SummarizedExperiment,numeric-method}
\alias{clusterCells,matrix,ANY-method}
\alias{clusterInfo,ClusterCells-method}
\alias{clusterLabels}
\alias{clusterLabels,ClusterCells,character-method}
\alias{clusterLabels,ClusterCells,missing-method}
\alias{clusterLabels,ClusterCells,numeric-method}
\alias{clusterLabels<-}
\alias{clusterLabels<-,ClusterCells,character-method}
\alias{clusterType,ClusterCells-method}
\alias{coClustering,ClusterCells-method}
\alias{dendrogram,ClusterCells-method}
\alias{nClusters}
\alias{nClusters,ClusterCells-method}
\alias{nFeatures}
\alias{nFeatures,ClusterCells-method}
\alias{nSamples}
\alias{nSamples,ClusterCells-method}
\alias{paramValues}
\alias{pipelineClusterDetails}
\alias{pipelineClusterTable}
\alias{pipelineClusters}
\alias{primaryCluster,ClusterCells-method}
\alias{primaryClusterIndex,ClusterCells-method}
\alias{primaryClusterIndex<-,ClusterCells,numeric-method}
\alias{removeClusters}
\alias{removeClusters,ClusterCells,character-method}
\alias{removeClusters,ClusterCells,numeric-method}
\alias{removeUnclustered,ClusterCells-method}
\alias{show,ClusterCells-method}
\alias{transform}
\alias{transformation}
\alias{transformation,ClusterCells-method}
\title{Class ClusterCells}
\usage{
clusterCells(se, clusters, ...)

\S4method{clusterCells}{matrix,ANY}(se, clusters, ...)

\S4method{clusterCells}{SummarizedExperiment,numeric}(se, clusters, ...)

\S4method{clusterCells}{SummarizedExperiment,character}(se, clusters, ...)

\S4method{clusterCells}{SummarizedExperiment,factor}(se, clusters, ...)

\S4method{clusterCells}{SummarizedExperiment,matrix}(se, clusters,
  transformation, clusterType = "User", clusterInfo = NULL)

clusterLabels(x, whichClusters)

nClusters(x)

nFeatures(x)

nSamples(x)

clusterLabels(object) <- value

pipelineClusters(x, iteration = 0)

pipelineClusterDetails(x)

pipelineClusterTable(x)

paramValues(x, type)

transformation(x)

transform(x, nPCADims = NA, nVarDims = NA, dimReduce = "none")

allClusters(x)

removeClusters(x, whichRemove, exactMatch = TRUE)

\S4method{show}{ClusterCells}(object)

\S4method{transformation}{ClusterCells}(x)

\S4method{clusterLabels}{ClusterCells,character}(object) <- value

\S4method{clusterLabels}{ClusterCells,numeric}(x, whichClusters)

\S4method{clusterLabels}{ClusterCells,character}(x, whichClusters = "all")

\S4method{clusterLabels}{ClusterCells,missing}(x, whichClusters)

\S4method{nClusters}{ClusterCells}(x)

\S4method{nFeatures}{ClusterCells}(x)

\S4method{nSamples}{ClusterCells}(x)

\S4method{allClusters}{ClusterCells}(x)

\S4method{primaryCluster}{ClusterCells}(x)

\S4method{primaryClusterIndex}{ClusterCells}(x)

\S4method{primaryClusterIndex}{ClusterCells,numeric}(object) <- value

\S4method{coClustering}{ClusterCells}(x)

\S4method{dendrogram}{ClusterCells}(x)

\S4method{clusterType}{ClusterCells}(x)

\S4method{removeClusters}{ClusterCells,character}(x, whichRemove,
  exactMatch = TRUE)

\S4method{removeClusters}{ClusterCells,numeric}(x, whichRemove)

\S4method{addClusters}{ClusterCells,matrix}(x, y, type = "User")

\S4method{addClusters}{ClusterCells,numeric}(x, y, type = "User")

\S4method{addClusters}{ClusterCells,ClusterCells}(x, y)

\S4method{removeUnclustered}{ClusterCells}(x)

\S4method{clusterInfo}{ClusterCells}(x)
}
\arguments{
\item{se}{a matrix or \code{SummarizedExperiment} containing the clustered
data.}

\item{transformation}{function. A function to transform the data before performing steps that assume normal-like (i.e. constant variance), such as the log}

\item{labels}{a vector with cluster labels.}
}
\value{
A \code{ClusterCells} object.
}
\description{
\code{ClusterCells} is a class that extends \code{SummarizedExperiment} and is
used to store the single-cell RNA-seq data and clustering information.

In addition to the slots of the \code{SummarizedExperiment} class, the
\code{ClusterCells} object has the following additional slots:

The constructor \code{clusterCells} creates an object of the class
\code{ClusterCells}. However, the typical way of creating these objects is
the result of a call to \code{clusterMany} or \code{clusterAll}.

Note that when subsetting the data, the co-clustering and dendrogram
information are lost.
}
\section{Slots}{

\describe{
\item{\code{transformation}}{function. Function to transform the data by when methods that assume normal-like data (e.g. log)}

\item{\code{clusterMatrix}}{matrix. A matrix giving the
integer-valued cluster ids for each sample. The rows of the matrix correspond to clusterings and columns to samples. 
The integer values are assigned
in the order that the clusters were found, if found by setting sequential=TRUE in clusterAll. "-1" indicates
the sample was not clustered.}

\item{\code{primaryIndex:}}{numeric. An index that specifies the primary set of
labels.}

\item{\code{clusterInfo:}}{list. A list with info about the clustering. 
If created from \code{\link{clusterAll}}, clusterInfo will include the
parameter used for the call, and the call itself. If \code{sequential = TRUE}
it will also include the following components.

\itemize{
\item{\code{clusterInfo}}{if sequential=TRUE and clusters were successfully
found, a matrix of information regarding the algorithm behavior for each
cluster (the starting and stopping K for each cluster, and the number of
iterations for each cluster).}
\item{\code{whyStop}}{if sequential=TRUE and clusters were successfully
found, a character string explaining what triggered the algorithm to stop.}
}}

\item{\code{clusterType:}}{character vector with the origin of each column of
clusterMatrix.}

\item{\code{dendrogram:}}{dendrogram. A dendrogram containing the cluster
relationship.}

\item{\code{coClustering:}}{matrix. A matrix with the cluster co-occurrence
information; this can either be based on subsampling or on co-clustering
across parameter sets (see \code{clusterMany}). The matrix is a square matrix with number of rows/columns equal to the number of samples.}
}}
\examples{

se <- matrix(data=rnorm(200), ncol=10)
labels <- gl(5, 2)

cc <- clusterCells(se, as.numeric(labels), transformation = function(x){x})

}

