% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllHelper.R
\docType{class}
\name{ClusterCells-class}
\alias{ClusterCells}
\alias{ClusterCells-class}
\alias{[,ClusterCells,ANY,ANY,ANY-method}
\alias{addClusters,ClusterCells,ClusterCells-method}
\alias{addClusters,ClusterCells,matrix-method}
\alias{addClusters,ClusterCells,numeric-method}
\alias{allClusters,ClusterCells-method}
\alias{clusterCells}
\alias{clusterType,ClusterCells-method}
\alias{coClustering,ClusterCells-method}
\alias{dendrogram,ClusterCells-method}
\alias{isLog,ClusterCells-method}
\alias{primaryCluster,ClusterCells-method}
\alias{primaryClusterIndex,ClusterCells-method}
\alias{primaryClusterIndex<-,ClusterCells,numeric-method}
\alias{removeUnclustered,ClusterCells-method}
\alias{show,ClusterCells-method}
\title{ClusterCells object and constructor}
\usage{
clusterCells(se, labels, isLog)

\S4method{[}{ClusterCells,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{ClusterCells}(object)

\S4method{isLog}{ClusterCells}(x)

\S4method{allClusters}{ClusterCells}(x)

\S4method{primaryCluster}{ClusterCells}(x)

\S4method{primaryClusterIndex}{ClusterCells}(x)

\S4method{primaryClusterIndex}{ClusterCells,numeric}(object) <- value

\S4method{coClustering}{ClusterCells}(x)

\S4method{dendrogram}{ClusterCells}(x)

\S4method{clusterType}{ClusterCells}(x)

\S4method{addClusters}{ClusterCells,matrix}(x, y)

\S4method{addClusters}{ClusterCells,numeric}(x, y)

\S4method{addClusters}{ClusterCells,ClusterCells}(x, y)

\S4method{removeUnclustered}{ClusterCells}(x)
}
\arguments{
\item{se}{a matrix or \code{SummarizedExperiment} containing the clustered
data.}

\item{labels}{a vector with cluster labels.}

\item{isLog}{logical. Whether the data are in log (TRUE) or linear (FALSE)
scale}
}
\value{
A \code{ClusterCells} object.
}
\description{
\code{ClusterCells} is a class that extends \code{SummarizedExperiment},
used to store the single-cell RNA-seq data and clustering information.
In addition to the slots of the \code{SummarizedExperiment} class, the
\code{ClusterCells} object has the following additional slots:
\itemize{
\item isLog: logical. Whether the data are in the linear or log scale.
\item clusterLabels: matrix. A matrix of cluster labels, useful for consensus
clustering.
\item primaryIndex: numeric. An index that specifies the primary set of
labels.
\item clusterInfo: list. A list with info about the clustering.
\itam clusterType: character vector with the origin of each column of
clusterLabels.
\item dendrogram: dendrogram. A dendrogram containing the cluster
relationship.
\item coClustering: matrix. A matrix with the cluster co-occurrence
information; this can either be based on subsampling or on co-clustering
across parameter sets (see \code{compareChoices}).
}
}
\details{
The constructor \code{clusterCells} creates an object of the class
\code{ClusterCells}. However, the typical way of creating these objects is
the result of a call to \code{compareChoices} or \code{clusterAll}.

Note that when subsetting the data, the co-clustering and dendrogram
information are lost.
}
\examples{

se <- matrix(data=rnorm(200), ncol=10)
labels <- gl(5, 2)

cc <- clusterCells(se, as.numeric(labels), isLog = TRUE)

}

