% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecateFunctions.R
\name{clusterExperiment-deprecated}
\alias{clusterExperiment-deprecated}
\alias{combineMany}
\alias{combineMany,ANY-method}
\title{Deprecated functions in package \sQuote{clusterExperiment}}
\usage{
\S4method{combineMany}{ANY}(x, ...)
}
\arguments{
\item{x}{any object}

\item{...}{additional arguments}
}
\description{
These functions are provided for compatibility with older versions
 of \sQuote{clusterExperiment} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
 \itemize{
   \item \code{combineMany}: \code{\link{makeConsensus}}
   \item \code{removeUnclustered}: \code{\link{removeUnassigned}}
 }
}
