% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelper.R
\docType{methods}
\name{ClusterExperiment-methods}
\alias{ClusterExperiment-methods}
\alias{clusterInfo,ClusterExperiment-method}
\alias{clusterLegend,ClusterExperiment-method}
\alias{clusterLegend<-,ClusterExperiment,list-method}
\alias{clusterMatrix,ClusterExperiment-method}
\alias{clusterMatrixNamed,ClusterExperiment-method}
\alias{clusterType,ClusterExperiment-method}
\alias{clusterType<-,ClusterExperiment,character-method}
\alias{coClustering,ClusterExperiment-method}
\alias{coClustering<-,ClusterExperiment,matrix-method}
\alias{nClusters,ClusterExperiment-method}
\alias{nFeatures,ClusterExperiment-method}
\alias{nSamples,ClusterExperiment-method}
\alias{orderSamples,ClusterExperiment-method}
\alias{orderSamples<-,ClusterExperiment,numeric-method}
\alias{primaryCluster,ClusterExperiment-method}
\alias{primaryClusterIndex,ClusterExperiment-method}
\alias{primaryClusterIndex<-,ClusterExperiment,numeric-method}
\alias{primaryClusterNamed,ClusterExperiment-method}
\alias{show,ClusterExperiment-method}
\alias{transformation,ClusterExperiment-method}
\title{Helper methods for the ClusterExperiment class}
\usage{
\S4method{[}{ClusterExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{ClusterExperiment}(object)

\S4method{clusterMatrixNamed}{ClusterExperiment}(x)

\S4method{primaryClusterNamed}{ClusterExperiment}(x)

\S4method{transformation}{ClusterExperiment}(x)

\S4method{nClusters}{ClusterExperiment}(x)

\S4method{nFeatures}{ClusterExperiment}(x)

\S4method{nSamples}{ClusterExperiment}(x)

\S4method{clusterMatrix}{ClusterExperiment}(x)

\S4method{primaryCluster}{ClusterExperiment}(x)

\S4method{primaryClusterIndex}{ClusterExperiment}(x)

\S4method{primaryClusterIndex}{ClusterExperiment,numeric}(object) <- value

\S4method{coClustering}{ClusterExperiment}(x)

\S4method{coClustering}{ClusterExperiment,matrix}(object) <- value

\S4method{clusterType}{ClusterExperiment}(x)

\S4method{clusterInfo}{ClusterExperiment}(x)

\S4method{clusterLegend}{ClusterExperiment}(x)

\S4method{clusterLegend}{ClusterExperiment,list}(object) <- value

\S4method{orderSamples}{ClusterExperiment}(x)

\S4method{orderSamples}{ClusterExperiment,numeric}(object) <- value

\S4method{clusterType}{ClusterExperiment,character}(object) <- value
}
\arguments{
\item{x}{a ClusterExperiment object.}

\item{object}{a ClusterExperiment object.}
}
\description{
This is a collection of helper methods for the ClusterExperiment class.
}
\details{
Note that when subsetting the data, the dendrogram information and
the co-clustering matrix are lost.

\code{clusterMatrixNamed} returns a matrix with cluster labels.

\code{primaryClusterNamed} returns the primary cluster (using cluster
labels).

\code{transformation} prints the function used to transform the data
prior to clustering.

\code{nClusters} returns the number of clusterings (i.e., ncol of
clusterMatrix).

\code{nFeatures} returns the number of features (same as `nrow`).

\code{nSamples} returns the number of samples (same as `ncol`).

\code{clusterMatrix} returns the matrix with all the clusterings.

\code{primaryCluster} returns the primary clustering (as numeric).

\code{primaryClusterIndex} returns/sets the primary clustering index
(i.e., which column of clusterMatrix corresponds to the primary clustering).

\code{coClustering} returns/sets the co-clustering matrix.

\code{clusterType} returns/sets the clusterType slot.

\code{clusterInfo} returns the clusterInfo slot.

\code{clusterLegend} returns/sets the clusterLegend slot.

\code{orderSamples} returns/sets the orderSamples slot.
}

