% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusterHclust.R
\name{clusterHclust}
\alias{clusterHclust}
\title{Make hierarchy of set of clusters}
\usage{
clusterHclust(dat, cl, full = TRUE, unassigned = c("remove", "outgroup",
  "cluster"), ...)
}
\arguments{
\item{dat}{data to define the mediods from. Samples on rows, features on columns}

\item{cl}{A vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.}

\item{full}{Logical as to whether to expand to include all samples as tips (useful for heatmap), or to have dendrogram just of the clusters (useful for getBestGenes)}

\item{unassigned}{what to do with with unassigned ("-1") cluster; only relevant if full=TRUE. Default is to remove those samples, so that only get heatmap of clustered observations. ``outgroup'' puts the unassigned into an outgroup on the dendrogram. ``cluster'' clusters them with the mediods.}
}
\description{
Makes a dendrogram of a set of clusters based on hclust on the mediods of the cluster.
}
\details{
If full=TRUE, then the expanded dendrogram is created by giving all of the samples the mediod of the cluster and applying hclust to it; if further unassigned=="cluster", then the expanded dendrogram is created by hclust of the expanded mediod data plus the original unclustered observations.
}

