% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareChoices.R
\name{compareChoices}
\alias{compareChoices}
\title{Create a matrix of clustering across values of k}
\usage{
compareChoices(data, ks = 2:15, alphas = c(0.1, 0.2, 0.3),
  findBestK = c(TRUE, FALSE), sequential = c(TRUE, FALSE),
  removeSil = c(TRUE, FALSE), subsample = c(TRUE, FALSE),
  clusterMethod = c("tight", "hierarchical", "pam"),
  subsampleClusterArgs = list(nstart = 10), clusterArgs = list(min.size =
  5), subsampleArgs = list(resamp.num = 50), seqArgs = list(beta = 0.9,
  k.min = 3), ncores = 1, random.seed = NULL, ...)
}
\arguments{
\item{data}{the data on which to run the clustering (samples in rows)}

\item{ks}{the range of k values (see details for meaning for different choices)}

\item{alphas}{values of alpha to be tried. Only used for subsampleClusterMethod either 'tight' or 'hierarchical'; otherwise (i.e. pam) alpha=0 for removal of negative silhouette values [can't adjust this at this time.]}

\item{findBestK}{values of findBestK to be tried (logical).}

\item{sequential}{values of sequential to be tried (logical)}

\item{removeSil}{values of removeSil to be tried (logical)}

\item{clusterMethod}{if subsampleClusterMethod!="none", method used in clustering of subsampled data; if ="none", method for clustering dist(x) [currently must be "pam]}

\item{subsampleClusterArgs}{list of arguments to be passed to clusterDMat}

\item{clusterArgs}{list of arguments to be passed to subsampleClustering's argument 'clusterArgs' that control clustering of subsampled data}

\item{subsampleArgs}{list of arguments to be passed to subsamplingClustering}

\item{ncores}{the number of threads}

\item{random.seed}{a value to set seed before each run of clusterAll (so that all of the runs are run on the same subsample of the data)}

\item{...}{to be passed on to mclapply (if ncores>1)}

\item{subsampleClusterMethod}{if "none", no subsampling is done, and clustering on dist(x) with clusterMethod technique. Otherwise, passed to clusterDMat option 'method' to indicate method of clustering of the co-occurrence matrix from subsampleClustering.}
}
\value{
A list with the following objects:
\itemize{
\item{\code{clMat}}{a matrix of with each row corresponding to a clustering and each column a sample.}
\item{\code{clusterInfo}}{a list with information regarding clustering result for those clusterings with sequential=TRUE}
}
}
\description{
Given a range of k's, this funciton will return a matrix with the clustering of the samples
across the range, which can be passed to \code{plotTracking} for visualization.
}
\details{
Currently there is no functionality to run over the different subsampleClusterMethods, in the sense that there is no way to reuse the same subsampling matrix and try different subsampleClusterMethods on it. This is because if sequential=TRUE, different subsampleClusterMethods will create different sets of data to subsample (and if sequential=FALSE, we have not implemented functionality for this reuse). Therefore, only one value of subsampleClusterMethod is allowed
}
\examples{
data(simData)
system.time(clusterTrack<-compareChoices(simData, ks=2:15, alphas=c(0.1,0.2,0.3), #'	findBestK=c(TRUE,FALSE),sequential=c(FALSE),subsample=c(FALSE),removeSil=c(TRUE),
	clusterMethod="pam",
	clusterArgs = list(min.size = 5,kRange=2:15),
	ncores=1,random.seed=48120))
}

