% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelpers.R
\docType{methods}
\name{convertClusterLegend}
\alias{convertClusterLegend}
\title{Convert clusterLegend into useful formats}
\usage{
\S4method{convertClusterLegend}{ClusterExperiment}(object,
  output = c("plotAndLegend", "aheatmapFormat", "matrixNames",
  "matrixColors"))
}
\arguments{
\item{object}{a \code{ClusterExperiment} object.}

\item{output}{character value, indicating desired type of conversion.}
}
\value{
If \code{output="plotAndLegend"}, \code{"convertClusterLegend"} will
  return a list that provides the necessary information to color samples
  according to cluster and create a legend for it:
 \itemize{
 \item{"colorVector"}{ A vector the same length as the number of samples,
 assigning a color to each cluster of the primaryCluster of the object.}
 \item{"legendNames"}{ A vector the length of the number of clusters of
 primaryCluster of the object giving the name of the cluster.}
 \item{"legendColors"}{ A vector the length of the number of clusters of
 primaryCluster of the object giving the color of the cluster.}
}

If \code{output="aheatmap"} a conversion of the clusterLegend to be
  in the format requested by \code{\link[NMF]{aheatmap}}. The column 'name'
  is used for the names and the column 'color' for the color of the clusters.

If \code{output="matrixNames"} or \code{"matrixColors"} a matrix the
  same dimension of \code{clusterMatrix(object)}, but with the cluster color
  or cluster name instead of the clusterIds, respectively.
}
\description{
Function for converting the information stored in the clusterLegend slot into
other useful formats.
}
\details{
convertClusterLegend pulls out information stored in the
  \code{clusterLegend} slot of the object and returns it in useful format.
}

