% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduce.R
\docType{methods}
\name{dimReduction,ClusterExperiment-method}
\alias{dimReduction,ClusterExperiment-method}
\title{General wrapper to perform dimensionality reduction}
\usage{
\S4method{dimReduction}{ClusterExperiment}(x, methods = "PCA", k)
}
\arguments{
\item{x}{a \code{SingleCellExperiment} object.}

\item{methods}{a vector of character strings that specify which methods to
use for dimensionality reduction.}

\item{k}{the number of dimensions.}
}
\description{
Given a \code{\link{ClusterExperiment}} object, it returns a copy of the same
object with the additional dimensionality reduced data in the
\code{dimReduce} slot.
}
\details{
Note that with this function one can perform several dimensionality reduction
techniques on the data at the same time. The results will be stored as a
\code{SimpleList} in the \code{reducedDim} slot.

For now, only PCA is implemented.
}
