% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getBestGenes}
\alias{getBestGenes}
\title{Function for finding best genes associated with clusters}
\usage{
getBestGenes(cl, dat, type = c("F", "Dendro", "Pairs", "OneAgainstAll"),
  dendro = NULL, pairMat = NULL, returnType = c("Table", "Index"),
  contrastAdj = c("All", "PerContrast", "AfterF"), voomCorrection = FALSE,
  ...)
}
\arguments{
\item{cl}{A numeric vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.}

\item{dat}{data for the test, samples in rows}

\item{type}{What type of test to do. `F' gives the omnibus F-statistic, `Dendro' traverses the given dendrogram and does contrasts of the samples in each side,  `Pairs' does pair-wise contrasts based on the pairs given in pairMat (if pairMat=NULL, does all pairwise), and `OneAgainstAll' compares each cluster to the average of all others.}

\item{dendro}{The dendrogram to traverse if type="Dendro". Note that this should be the dendrogram of the clusters, not of the individual samples.}

\item{pairMat}{matrix giving the pairs of clusters for which to do pair-wise contrasts (must match to elements of cl). If NULL, will do all pairwise of the clusters in \code{cl} (excluding "-1" categories). Each row is a pair to be compared and must match the names of the clusters in the vector \code{cl}.}

\item{returnType}{Whether to return the index of genes, or the full table given by topTable or topTableF.}

\item{contrastAdj}{What type of FDR correction to do for contrasts tests (i.e. if type='Dendro' or 'Pairs').}

\item{voomCorrection}{Whether to perform voom correction to data, e.g. if input data matrix is counts/TPM/FPKM. If input to dat consists of counts/TPM/FPKM, this argument should be set to TRUE. Otherwise, dat should be something like log of counts (which is not preferable for count data) or some other kind of similar input data that does not need a variance stabilization (e.g. microarray data). Currently the default is set to FALSE, simply because the voomCorrection has not been heavily tested. But TRUE with dat being counts/TPM/FPKM really should be the default for RNA-Seq data.}

\item{...}{options to pass to \code{\link{topTable}} or \code{\link{topTableF}} (see \code{\link{limma}} package)}
}
\value{
A data.frame in the same format as \code{\link{topTable}}, except for the following additional or changed columns:
\itemize{
\item{\code{Gene}}{This is the column called 'ProbeID' by \code{\link{topTable}}}
\item{\code{IndexInOriginal}}{Gives the index of the gene to the original input dataset, \code{dat}}
\item{\code{Contrast}}{The contrast that the results corresponds to (if applicable, depends on \code{type} argument)}
\item{\code{ContrastName}}{The name of the contrast that the results corresponds to. For dendrogram searches, this will be the node of the tree of the dendrogram.}
}
}
\description{
Calls limma on input data to determine genes most associated with found clusters (either pairwise comparisons, or following a tree that clusters the clusters).
}
\details{
getBestGenes returns best genes corresponding to a cluster assignment. It uses limma to fit the models, and limma's functions topTable or topTableF to find the best genes.  See the options of these functions to put better control on what gets returned (e.g. only if significant, only if log-fc is above a certain amount, etc.). In particular, set 'number=' to define how many significant genes to return (where number is per contrast for the 'Pairs' or 'Dendro' option)
}
\examples{
data(simData)
#create a clustering, for 8 clusters (truth was 4)
cl<-clusterAll(simData,clusterFunction="pam",subsample=FALSE,
sequential=FALSE, clusterDArgs=list(k=8))$cl
#basic F test, return all, even if not significant:
testF<-getBestGenes(cl,simData,type="F",number=nrow(simData),
voomCorrection=FALSE)
#Do all pairwise, only return significant, try different adjustments:
pairsPerC<-getBestGenes(cl,simData,type="Pairs",contrastAdj="PerContrast",
p.value=0.05,voomCorrection=FALSE)
pairsAfterF<-getBestGenes(cl,simData,type="Pairs",contrastAdj="AfterF",
p.value=0.05,voomCorrection=FALSE)
pairsAll<-getBestGenes(cl,simData,type="Pairs",contrastAdj="All",
p.value=0.05,voomCorrection=FALSE)
#not useful for this silly example, but could look at overlap with Venn
allGenes<-paste("Row",1:nrow(simData),sep="")
if(require(limma)){
	vennC<-vennCounts(cbind(PerContrast=allGenes\%in\%pairsPerC$Gene,
AllJoint=allGenes\%in\%pairsAll$Gene,FHier=allGenes\%in\% pairsAfterF$Gene))
	vennDiagram(vennC,main="FDR Overlap")
}
#Do one cluster against all others
oneAll<-getBestGenes(cl,simData,type="OneAgainstAll",contrastAdj="All",p.value=0.05)
#Do dendrogram testing
hcl<-clusterHclust(dat=simData,cl,full=FALSE)
allDendro<-getBestGenes(cl=cl,dat=simData,type="Dendro",dendro=hcl,returnType=c("Table"),
contrastAdj=c("All"),number=ncol(simData),p.value=0.05)

# do DE on counts using voom
# compare results to if used simData instead (not on count scale).
# Again, not relevant for this silly example, but basic principle useful
testFVoom<-getBestGenes(cl,simCount,type="F",number=nrow(simData),voomCorrection=TRUE)
plot(testF$P.Value[order(testF$Index)],testFVoom$P.Value[order(testFVoom$Index)],log="xy")
}

