% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracking.R
\name{makeAnnoColor}
\alias{makeAnnoColor}
\title{Make anno for input to dualHeatmap from output of plotTracking}
\usage{
makeAnnoColor(trackingOut)
}
\arguments{
\item{trackingOut}{the output from \code{\link{plotTracking}}}
}
\description{
Make anno for input to dualHeatmap from output of plotTracking
}

