% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFunction.R
\docType{methods}
\name{makeDimReduce}
\alias{makeDimReduce}
\alias{makeDimReduce,matrix-method}
\alias{makeDimReduce,SummarizedExperiment-method}
\alias{makeDimReduce,ClusterExperiment-method}
\alias{listBuiltInDimReduce}
\alias{makeFilterStats,matrix-method}
\title{Calculate dimensionality reduction of data}
\usage{
\S4method{makeDimReduce}{SingleCellExperiment}(object, dimReduce = "PCA",
  maxDims = 500, transFun = NULL, isCount = FALSE)

\S4method{makeDimReduce}{matrix}(object, ...)

\S4method{makeDimReduce}{SummarizedExperiment}(object, ...)

\S4method{makeDimReduce}{ClusterExperiment}(object, ...)

listBuiltInDimReduce()

\S4method{makeFilterStats}{matrix}(object, ...)
}
\arguments{
\item{object}{input to use for the data for dimensionality reduction. Can be matrix, SummarizedExperiment, SingleCellExperiment, or ClusterExperiment object}

\item{dimReduce}{a vector of character values indicating the methods of dimensionality reduction to be performed. Currently only "PCA" is implemented.}

\item{maxDims}{Numeric vector of integer giving the number of PC dimensions to calculate. 
\code{maxDims} can also take values between (0,1) to indicate keeping the
number of PCA dimensions necessary to account for that proportion of the
variance. \code{maxDims} should be of same length as \code{dimReduce}, indicating the number of dimensions to keep for each method (if \code{maxDims} is of length 1, the same number of dimensions will be used for each).}

\item{transFun}{a function that will be used to transform the data before performing dimensionality reduction. If \code{object} is a \code{ClusterExperiment} object, then the value of \code{transformation} slot will be used and the user cannot give a value for \code{transFun}.}

\item{isCount}{logical. If \code{transFun} is not given, and \code{isCount=TRUE}, then the \code{transFun} will be assumed to be \code{log(x+1)}.}
}
\value{
a SingleCellExperiment object with the indicated diminsionality reduction methods stored in the \code{reduceDims} slot.

A \code{\link{SingleCellExperiment}} containing the dimensionality reduction in the corresponding slots with names corresponding to the name given in \code{dimReduce}.
}
\description{
A Function for performing and storing common dimensionality reduction techniques
}
\details{
The PCA method uses either \code{prcomp} from the \code{stats} package or  \code{svds} from the \code{RSpectra} package to perform PCA. Both are called on \code{t(assay(x))} with \code{center=TRUE} and \code{scale=TRUE} (i.e. the feature are centered and scaled), so that
  it is performing PCA on the correlation matrix of the features.
}
\examples{
mat <- matrix(data=rnorm(200), ncol=10)
mat[1,1] <- -1 #force a negative value
labels <- gl(5, 2)
cc <- ClusterExperiment(mat, as.numeric(labels), transformation =
function(x){x^2}) #define transformation as x^2
#will transform data based on saved transformation
x <- makeDimReduce(cc, dimReduce="PCA", nPCADims=3)
}
