% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduce.R
\docType{methods}
\name{pca,ClusterExperiment-method}
\alias{pca,ClusterExperiment-method}
\title{Wrapper to perform PCA on a SingleCellExperiment}
\usage{
\S4method{pca}{ClusterExperiment}(x, k, center = TRUE, scale = FALSE)
}
\arguments{
\item{x}{a \code{SingleCellExperiment} object.}

\item{k}{the number of principal components.}

\item{center}{a logical value indicating whether the variables should be
shifted to be zero centered.}

\item{scale}{a logical value indicating whether the variables should be
scaled to have unit variance.}
}
\description{
Given a \code{\link{ClusterExperiment}} object, it
returns a copy of the same object with the principal components added in the
\code{dimReduce} slot with the name "PCA".
}
\details{
This method uses the \code{\link[RSpectra]{svds}} function which is faster
than \code{svd} when \code{k} is small.
}
