% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelper.R
\docType{methods}
\name{pipelineClusterDetails,ClusterExperiment-method}
\alias{pipelineClusterDetails}
\alias{pipelineClusterDetails,ClusterExperiment-method}
\alias{pipelineClusterTable}
\alias{pipelineClusterTable,ClusterExperiment-method}
\alias{pipelineClusters}
\title{Methods for pipeline clusters}
\usage{
\S4method{pipelineClusterDetails}{ClusterExperiment}(x)

\S4method{pipelineClusterTable}{ClusterExperiment}(x)

\S4method{pipelineClusters}{ClusterExperiment}(x, iteration = 0)
}
\description{
The main pipeline of the package is made of \code{\link{clusterMany}},
\code{\link{combineMany}}, and \code{\link{mergeClusters}}.
The clusterings from these functions (and not those obtained in a different
way) can be obtained with the functions documented here.
}

