% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusters.R
\docType{methods}
\name{plotClustersWorkflow,ClusterExperiment-method}
\alias{plotClustersWorkflow,ClusterExperiment-method}
\alias{plotClustersWorkflow}
\title{A plot of clusterings specific for clusterMany and workflow
  visualization}
\usage{
\S4method{plotClustersWorkflow}{ClusterExperiment}(object,
  whichClusterMany = NULL, whichResults = c("mergeClusters", "combineMany"),
  nBlankRows = ceiling(nClusters(object) * 0.05),
  nSizeResult = ceiling(nClusters(object) * 0.02), clusterManyLabels = TRUE,
  resultLabels = TRUE, sortBy = c("results", "clusterMany"),
  resultsOnTop = TRUE, ...)
}
\arguments{
\item{object}{A \code{ClusterExperiment} object on which
\code{\link{clusterMany}} has been run}

\item{whichClusterMany}{numeric indices of which of the clusterMany
clusterings to plot (if NULL, defaults to all)}

\item{whichResults}{which clusterings to use as the results.}

\item{nBlankRows}{the number of blank (i.e. white) rows to add between the
clusterMany clusterings and the results}

\item{nSizeResult}{the number of rows each result clustering should take up}

\item{clusterManyLabels}{logical, whether to plot the labels for the results
of clusterMany}

\item{sortBy}{how to align the clusters. If "results" then the results are in
the top of the alignment done by plotClusters. If "clusterMany", then the
clusterMany results are in the top. (Note this does not determine where
they will be plotted, but how they are ordered in the aligning step done by
\code{plotClusters})}

\item{resultsOnTop}{logical. Whether the results should be plotted on the top
of clusterMany results or the bottom.}
}
\description{
A realization of \code{\link{plotClusters}} call specific to
  separating out the results of \code{clusterMany} and other clustering
  results.
}
\seealso{
\code{\link{plotClusters}}, \code{\link{clusterMany}}
}
