% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContrastHeatmap.R
\docType{methods}
\name{plotContrastHeatmap,ClusterExperiment-method}
\alias{plotContrastHeatmap,ClusterExperiment-method}
\alias{plotContrastHeatmap}
\title{Plot heatmaps showing significant genes per contrast}
\usage{
\S4method{plotContrastHeatmap}{ClusterExperiment}(object, signifTable, ...)
}
\arguments{
\item{object}{ClusterExperiment object on which biomarkers were found}

\item{signifTable}{A \code{data.frame} in format of the result of \code{\link{getBestFeatures}}. It must minimally contain columns 'Contrast' and 'IndexInOriginal' giving the grouping and original index of the features in the \code{assay(object)}}

\item{...}{Arguments passed to \code{\link{plotHeatmap}}}
}
\description{
Plots a heatmap of the data, with the genes grouped based on the contrast for which they were significant.
}
\examples{
data(simData)

cl <- clusterSingle(simData, subsample=FALSE,
sequential=FALSE, mainClusterArgs=list(clusterFunction="pam", clusterArgs=list(k=8)))

#Do all pairwise, only return significant, try different adjustments:
pairsPerC <- getBestFeatures(cl, contrastType="Pairs", number=5,
p.value=0.05, isCount=FALSE)
plotContrastHeatmap(cl,pairsPerC)
}
\seealso{
\code{\link{plotHeatmap}}, \code{\link{makeBlankData}}, \code{\link{getBestFeatures}}
}
