% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureBoxplot.R
\docType{methods}
\name{plotFeatureBoxplot}
\alias{plotFeatureBoxplot}
\alias{plotFeatureBoxplot,ClusterExperiment,character,ANY-method}
\alias{plotFeatureBoxplot,ClusterExperiment,missing,ANY-method}
\alias{plotFeatureBoxplot,ClusterExperiment,numeric,character-method}
\alias{plotFeatureBoxplot,ClusterExperiment,numeric,numeric-method}
\title{Plot boxplot of feature values by cluster}
\usage{
\S4method{plotFeatureBoxplot}{ClusterExperiment,character,ANY}(object,
  whichCluster, feature, ...)

\S4method{plotFeatureBoxplot}{ClusterExperiment,missing,ANY}(object,
  whichCluster, feature, ...)

\S4method{plotFeatureBoxplot}{ClusterExperiment,numeric,character}(object,
  whichCluster, feature, ...)

\S4method{plotFeatureBoxplot}{ClusterExperiment,numeric,numeric}(object,
  whichCluster, feature, plotUnassigned = FALSE, unassignedColor = NULL,
  missingColor = NULL, main = NULL, whichAssay = 1, ...)
}
\arguments{
\item{object}{a ClusterExperiment object}

\item{whichCluster}{which clusters to show on the plot}

\item{feature}{identification of feature to plot, either row name or index}

\item{...}{arguments passed to \code{\link{boxplot}}}

\item{plotUnassigned}{whether to plot the unassigned samples as a cluster
(either -1 or -2)}

\item{unassignedColor}{If not NULL, should be character value giving the
color for unassigned (-2) samples (overrides \code{clusterLegend}) default.}

\item{missingColor}{If not NULL, should be character value giving the color
for missing (-2) samples (overrides \code{clusterLegend}) default.}

\item{main}{title of plot. If NULL, given default title.}

\item{whichAssay}{numeric or character specifying which assay to use. See
\code{\link[SummarizedExperiment]{assay}} for details.}
}
\value{
A plot is created. The output of boxplot is returned
}
\description{
Plot a boxplot of the (transformed) values for a particular
  gene, separated by cluster
}
\examples{
#clustering using pam: try using different dimensions of pca and different k
data(simData)

cl <- clusterMany(simData, nReducedDims=c(5, 10, 50), reducedDim="PCA",
clusterFunction="pam", ks=2:4, findBestK=c(TRUE,FALSE),
removeSil=c(TRUE,FALSE))
clusterLegend(cl)[[1]][,"name"]<-letters[1:nClusters(cl,ignoreUnassigned = FALSE)[1]]
plotFeatureBoxplot(cl,feature=1)
}
\seealso{
\code{\link{boxplot}}
}
