% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureScatter.R
\docType{methods}
\name{plotFeatureScatter}
\alias{plotFeatureScatter}
\alias{plotFeatureScatter,ClusterExperiment-method}
\title{Plot scatter plot of feature values colored by cluster}
\usage{
\S4method{plotFeatureScatter}{ClusterExperiment,character}(object, features,
  ...)
}
\arguments{
\item{object}{a ClusterExperiment object}

\item{...}{arguments passed to \code{\link{boxplot}}}
}
\description{
Plot a scatter plot of the (transformed) values for a set of
  gene expression values, colored by cluster
}
