% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTableClusters.R
\docType{methods}
\name{plotTableClusters}
\alias{plotTableClusters}
\alias{plotTableClusters,table-method}
\alias{tableClusters,ClusterExperiment,character-method}
\alias{tableClusters}
\alias{tableClusters,ClusterExperiment,missing-method}
\alias{tableClusters,ClusterExperiment,numeric-method}
\title{Plot heatmap of cross-tabs of 2 clusterings}
\usage{
\S4method{plotTableClusters}{ClusterExperiment}(object, whichClusters,
  ignoreUnassigned = FALSE, margin = NA, ...)

\S4method{plotTableClusters}{table}(object, clusterLegend = NULL,
  cluster = FALSE, ...)

\S4method{tableClusters}{ClusterExperiment,character}(object, whichClusters,
  ...)

\S4method{tableClusters}{ClusterExperiment,missing}(object, whichClusters, ...)

\S4method{tableClusters}{ClusterExperiment,numeric}(object, whichClusters,
  useNames = TRUE, ...)
}
\arguments{
\item{object}{ClusterExperiment object (or matrix with table result)}

\item{ignoreUnassigned}{logical as to whether to ignore unassigned clusters in the 
plotting. This means they will also be ignored in the calculations of the proportions 
(if \code{margin} not NA).}

\item{margin}{if NA, the counts from \code{tableClusters} will be plotted. Otherwise, 
\code{\link[package=base]{prop.table}} will be called and the argument \code{\margin} #' will be passed to \code{prop.table} to determine how proportions should be calculated.}

\item{cluster}{logical, whether to cluster the rows and columns of the table. Passed
to arguments \code{clusterFeatures} AND \code{clusterSamples} of \code{plotHeatmap}.}

\item{useNames}{for \code{tableClusters}, whether the output should be tabled
with names (\code{useNames=TRUE}) or ids (\code{useNames=FALSE})}
}
\description{
Plot heatmap of cross-tabs of two clusterings
}
\seealso{
\code{\link[package=base]{prop.table}}

\code{\link{plotHeatmap}}
}
