% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCreation.R
\docType{data}
\name{rsecFluidigm}
\alias{rsecFluidigm}
\alias{makeRsecFluidigmObject}
\alias{checkRsecFluidigmObject}
\title{RSEC run for vignette}
\format{ClusterExperiment object, the result of running \code{\link{RSEC}} on
fluidigm data described in vignette and available in the \code{scRNAseq}
package.}
\usage{
makeRsecFluidigmObject(se)

checkRsecFluidigmObject(object)
}
\description{
RSEC run for vignette
}
\examples{
# see code used create rsecFluidigm 
# (print out the function)
makeRsecFluidigmObject
#code actualy run to create rsecFluidigm:
\dontrun{
library(clusterExperiment)
data(fluidigmData)
data(fluidigmColData)
se<-SummarizedExperiment(assays=fluidigmData,
       colData=fluidigmColData)
rsecFluidigm<-makeRsecFluidigmObject(se)
checkRsecFluidigmObject(rsecFluidigm)
devtools::use_data(rsecFluidigm,overwrite=FALSE)
}
}
\seealso{
\code{\link[scRNAseq]{fluidigm}}
}
\author{
Elizabeth Purdom \email{epurdom@stat.berkeley.edu}
}
\keyword{data}
