% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{search_pairs}
\alias{search_pairs}
\title{Search pairs of samples that co-cluster across subsamples}
\usage{
search_pairs(clusterings)
}
\arguments{
\item{clusterings}{a matrix with the cluster labels}
}
\value{
A matrix with the co-clusters, but only the lower triangle
is populated.
}
\description{
Assume that our input is a matrix, with N columns and B rows
(the number of subsamples), storing integers -- the cluster labels.
}
