% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorDefinitions.R
\docType{data}
\name{showHeatmapPalettes}
\alias{seqPal1}
\alias{seqPal2}
\alias{seqPal3}
\alias{seqPal4}
\alias{seqPal5}
\alias{showHeatmapPalettes}
\title{Set of colors useful for heatmap gradients}
\format{An object of class \code{character} of length 16.}
\usage{
showHeatmapPalettes()

seqPal5

seqPal2

seqPal3

seqPal4

seqPal1
}
\description{
Set of colors useful for heatmap gradients
}
\details{
seqPal1-seqPal4 are palettes for the heatmap. showHeatmapPalettes() will show you
these palettes.
}
\examples{
#show the palette colors
showHeatmapPalettes()
#compare the palettes on heatmap
data(simData)
data(simData)
cl<-clusterSingle(simData,clusterFunction="pam",subsample=FALSE,
sequential=FALSE, clusterDArgs=list(k=8))$cl
par(mfrow=c(2,3))
plotHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal1,main="seqPal1")
plotHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal2,main="seqPal2")
plotHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal3,main="seqPal3")
plotHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal4,main="seqPal4")
plotHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal5,main="seqPal5")

}
\keyword{datasets}

