% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCreation.R
\name{simFunctions}
\alias{simFunctions}
\alias{simSEObject}
\title{Functions for creating simulated objects for testing/examples}
\usage{
simSEObject(mat)
}
\description{
Functions for creating simulated objects for testing/examples
}
\examples{
mat <- matrix(data=rnorm(20*15), ncol=15)
mat[1,1]<- -1 #force a negative value
colnames(mat)<-paste("Sample",1:ncol(mat))
rownames(mat)<-paste("Gene",1:nrow(mat))
se<-simSEObject(mat)
se
}
\author{
Elizabeth Purdom \email{epurdom@stat.berkeley.edu}
}
