% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SingleCellFilter-class}
\alias{SingleCellFilter-class}
\alias{SingleCellFilter}
\title{Class SingleCellFilter}
\description{
\code{SingleCellFilter} is a class that extends
\code{SingleCellExperiment} and is used to store filterStats on the data

In addition to the slots of the \code{SingleCellExperiment}
class, the \code{SingleCellFilter} object has the additional slots described
in the Slots section.
}
\details{
\code{filterStats}
}
\section{Slots}{

\describe{
\item{\code{filterStats}}{a numeric matrix where each column gives a summary statistics 
calculated per gene that can be used for filtering of genes.}
}}

