% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFilter.R
\docType{methods}
\name{SingleCellFilter-methods}
\alias{SingleCellFilter-methods}
\alias{SingleCellFilter}
\alias{filterStats,SingleCellFilter-method}
\alias{filterStat}
\alias{filterStat,SingleCellFilter,character-method}
\alias{filterNames}
\alias{filterNames,SingleCellFilter-method}
\alias{filterData}
\alias{filterData,SingleCellFilter-method}
\title{Helper methods for the SingleCellFilter class}
\usage{
filterStats(object, ...)

\S4method{filterStats}{SingleCellFilter}(object)

filterStat(object, type, ...)

\S4method{filterStat}{SingleCellFilter,character}(object, type)

filterNames(object, ...)

\S4method{filterNames}{SingleCellFilter}(object)

filterData(object, ...)

\S4method{filterData}{SingleCellFilter}(object, type, cutoff, percentile,
  absolute = FALSE, keepLarge = TRUE)
}
\arguments{
\item{object}{a SingleCellFilter object}

\item{type}{character indicating which filter statistic to use. Must match a single filter statistic in \code{object}}

\item{cutoff}{numeric. A value at which to filter the rows (genes) for the test statistic}

\item{percentile}{numeric. Either a number between 0,1 indicating what percentage of the rows (genes) to keep or an integer value indicated the number of rows (genes) to keep}

\item{absolute}{whether to take the absolute value of the filter statistic}

\item{keepLarge}{logical whether to keep rows (genes) with large values of the test statistic or small values of the test statistic.}
}
\description{
This is a collection of helper methods for the SingleCellFilter class.
}
