% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClusters.R
\docType{methods}
\name{addClusterings,ClusterExperiment,matrix-method}
\alias{addClusterings,ClusterExperiment,matrix-method}
\alias{addClusterings}
\alias{removeClusterings}
\alias{addClusterings,ClusterExperiment,ClusterExperiment-method}
\alias{addClusterings,ClusterExperiment,numeric-method}
\alias{removeClusterings,ClusterExperiment,character-method}
\alias{removeClusterings,ClusterExperiment,numeric-method}
\alias{removeUnclustered,ClusterExperiment-method}
\alias{removeUnclustered}
\alias{unassignSamples,ClusterExperiment,numeric-method}
\alias{unassignSamples}
\alias{unassignSamples,ClusterExperiment,character-method}
\title{Functions to add/remove clusters to ClusterExperiment}
\usage{
\S4method{addClusterings}{ClusterExperiment,matrix}(x, y,
  clusterTypes = "User")

\S4method{addClusterings}{ClusterExperiment,ClusterExperiment}(x, y)

\S4method{addClusterings}{ClusterExperiment,numeric}(x, y,
  clusterLabel = NULL, ...)

\S4method{removeClusterings}{ClusterExperiment,character}(x, whichClusters, ...)

\S4method{removeClusterings}{ClusterExperiment,numeric}(x, whichClusters)

\S4method{removeUnclustered}{ClusterExperiment}(x)

\S4method{unassignSamples}{ClusterExperiment,numeric}(x, whichClusters,
  clustersToRemove)

\S4method{unassignSamples}{ClusterExperiment,character}(x, whichClusters, ...)
}
\arguments{
\item{x}{a ClusterExperiment object.}

\item{y}{additional clusters to add to x. Can be a ClusterExperiment object
or a matrix/vector of clusters.}

\item{clusterTypes}{a string describing the nature of the clustering. The 
values `clusterSingle`, `clusterMany`, `mergeClusters`, `combineMany` are 
reserved for the clustering coming from the package workflow and should not 
be used when creating a new object with the constructor.}

\item{clusterLabel}{label(s) for the clusters being added.}

\item{...}{Passed to signature \code{ClusterExperiment,matrix}.}

\item{whichClusters}{optional argument that can be either numeric or
character value. If numeric, gives the indices of the \code{clusterMatrix}
to return; this can also be used to defined an ordering for the
clusterings. \code{whichClusters} can be a character value identifying the 
\code{clusterTypes} to be used, or if not matching \code{clusterTypes} then
\code{clusterLabels}; alternatively \code{whichClusters} can be either 
'all' or 'workflow' to indicate choosing all clusters or choosing all 
\code{\link{workflowClusters}}. If missing, the entire matrix of all
clusterings is returned.}
}
\value{
A \code{\link{ClusterExperiment}} object with the added clusters.

\code{removeClusterings} returns a \code{ClusterExperiment} object, 
 unless all clusters are removed, in which case it returns a
 \code{\link{SingleCellExperiment}} object.
}
\description{
These functions are used to add or remove clusters to a
\code{\link{ClusterExperiment}} object.
}
\details{
addClusterings adds y to x, and is thus not symmetric in the two 
  arguments. In particular, the \code{primaryCluster}, all of the dendrogram
  information, \code{coClustering}, and \code{orderSamples} are all kept from
  the x object, even if y is a ClusterExperiment.

\code{removeClusterings} removes the clusters given by
 \code{whichClusters}. If the
 \code{primaryCluster} is one of the clusters removed, the
 \code{primaryClusterIndex} is set to 1 and the dendrogram and coclustering
 matrix are discarded and orderSamples is set to \code{1:NCOL(x)}.

\code{removeUnclustered} removes all samples that are unclustered
  (i.e. -1 or -2 assignment) in the \code{primaryCluster} of x (so they may
  be unclustered in other clusters found in \code{clusterMatrix(x)}).

\code{unassignSamples} unassigns samples in \code{clustersToRemove} and assigns them to -1 (unassigned)
}
\examples{
data(simData)

cl1 <- clusterSingle(simData, subsample=FALSE,
sequential=FALSE, mainClusterArgs=list(clusterArgs=list(k=3), clusterFunction="pam"))
cl2 <- clusterSingle(simData, subsample=FALSE,
sequential=FALSE, mainClusterArgs=list(clusterArgs=list(k=3), clusterFunction="pam"))

addClusterings(cl1, cl2)
}
