% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/addClusters.R
\docType{methods}
\name{addClusters}
\alias{addClusters}
\alias{addClusters,ClusterExperiment,ClusterExperiment-method}
\alias{addClusters,ClusterExperiment,matrix-method}
\alias{addClusters,ClusterExperiment,numeric-method}
\alias{removeClusters}
\alias{removeClusters,ClusterExperiment,character-method}
\alias{removeClusters,ClusterExperiment,numeric-method}
\alias{removeUnclustered}
\alias{removeUnclustered,ClusterExperiment-method}
\title{Functions to add/remove clusters to ClusterExperiment}
\usage{
addClusters(x, y, ...)

removeClusters(x, whichRemove, exactMatch = TRUE)

removeUnclustered(x)

\S4method{addClusters}{ClusterExperiment,matrix}(x, y, clusterType = "User")

\S4method{addClusters}{ClusterExperiment,ClusterExperiment}(x, y)

\S4method{addClusters}{ClusterExperiment,numeric}(x, y, ...)

\S4method{removeClusters}{ClusterExperiment,character}(x, whichRemove,
  exactMatch = TRUE)

\S4method{removeClusters}{ClusterExperiment,numeric}(x, whichRemove)

\S4method{removeUnclustered}{ClusterExperiment}(x)
}
\arguments{
\item{x}{a ClusterExperiment object.}

\item{y}{additional clusters to add to x. Can be a ClusterExperiment object
or a matrix/vector of clusters.}

\item{whichRemove}{which clusters to remove. Can be numeric or character. If
numeric, must give indices of \code{clusterMatrix(x)} to remove. If
character, should match a \code{clusterType} of x.}

\item{exactMatch}{logical. Whether \code{whichRemove} must exactly match a
value of \code{clusterType(x)}. Only relevant if whichRemove is character.}
}
\description{
These functions are used to add or remove clusters to a
\code{\link{ClusterExperiment}} object.
}
\details{
addClusters adds y to x, and is thus not symmetric in the two
  arguments. In particular, the \code{primaryCluster} and all of its
  supporting information (dendrogram, coClustering, and orderSamples) are all
  kept from the x object, even if y is a ClusterExperiment.

\code{removeClusters} removes the clusters given by
 \code{whichRemove}. If all clusters are implied, then returns a
 \code{\link[SummarizedExperiment]{SummarizedExperiment}} object. If the
 \code{primaryCluster} is one of the clusters removed, the
 \code{primaryClusterIndex} is set to 1 and the dendrogram and cooccurance
 matrix are discarded and orderSamples is set to \code{1:NCOL(x)}.

\code{removeUnclustered} removes all samples that are unclustered
  (i.e. -1 or -2 assignment) in the \code{primaryCluster} of x (so they may
  be unclustered in other clusters found in \code{clusterMatrix(x)}).
}
\examples{
data(simData)

cl1 <- clusterSingle(simData, clusterFunction="pam", subsample=FALSE,
sequential=FALSE, clusterDArgs=list(k=3))

cl2 <- clusterSingle(simData, clusterFunction="pam", subsample=FALSE,
sequential=FALSE, clusterDArgs=list(k=5))

addClusters(cl1, cl2)
}

