% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorDefinitions.R
\docType{data}
\name{bigPalette}
\alias{bigPalette}
\alias{showThisPal}
\title{Large palette of colors}
\format{\preformatted{ chr [1:60] "#1F78B4" "#33A02C" "#E31A1C" "#FF7F00" ...
}}
\usage{
bigPalette

showThisPal()
}
\description{
Large palette of colors
}
\details{
\code{bigPalette} is a long palette of colors (length 62) used by \code{\link{plotTracking}} and
accompanying functions. \code{showThisPal} creates plot that gives index of each color in bigPalette.
}
\keyword{datasets}

