% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtInClusterFunctions.R
\docType{methods}
\name{listBuiltInFunctions}
\alias{listBuiltInFunctions}
\alias{getBuiltInFunction,character-method}
\alias{getBuiltInFunction}
\alias{listBuiltInTypeK}
\alias{listBuiltInType01}
\title{Built in ClusterFunction options}
\usage{
listBuiltInFunctions()

\S4method{getBuiltInFunction}{character}(object)

listBuiltInTypeK(object)

listBuiltInType01(object)
}
\arguments{
\item{object}{unneeded argument.}
}
\description{
Documents the built-in clustering options that are available in
  the clusterExperiment package.
}
\details{
\code{builtInClusteringFunctions} will return the character names of
  the built-in clustering functions available.

\code{getBuiltInFunction} will return the
  \code{ClusterFunction} object of a character value that corresponds to a
  built-in function.

\code{\link{algorithmType}} and \code{\link{inputType}} will 
return the \code{algorithmType} and \code{inputType} of the
  built-in clusterFunction corresponding to the character value.

\code{listBuiltInTypeK} returns the names of the built-in functions
  that have type 'K'

\code{listBuiltInType01} returns the names of the built-in functions
  that have type '01'
}
\examples{
listBuiltInFunctions()
algorithmType(c("kmeans","pam","hierarchical01"))
inputType(c("kmeans","pam","hierarchical01"))
listBuiltInTypeK()
listBuiltInType01()
}
