% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAll.R
\name{clusterAll}
\alias{clusterAll}
\title{General wrap-around for all clustering methods we're trying.}
\usage{
clusterAll(x, subsample = TRUE, sequential = FALSE,
  clusterFunction = c("tight", "hierarchical", "pam", "kmeans"),
  clusterDArgs = NULL, subsampleArgs = NULL, seqArgs = NULL)
}
\arguments{
\item{x}{the data on which to run the clustering (samples in rows)}

\item{subsample}{logical as to whether to subsample via \code{\link{subsampleClustering}} to get the distance matrix at each iteration; otherwise the distance matrix is dist(x)}

\item{sequential}{logical whether to use the sequential strategy.}

\item{clusterFunction}{passed to \code{\link{clusterD}} option 'clusterFunction' to indicate method of clustering, see \code{\link{clusterD}}}

\item{clusterDArgs}{list of additional arguments to be passed to \code{\link{clusterD}}}

\item{subsampleArgs}{list of arguments to be passed to \code{\link{subsampleClustering}}}

\item{seqArgs}{list of additional arguments to be passed to \code{\link{seqCluster}}}
}
\value{
A list with values 
\itemize{
\item{\code{clustering}}{a vector of length equal to nrows(x) giving the integer-valued cluster ids for each sample. The integer values are assigned in the order that the clusters were found, if sequential=TRUE. "-1" indicates the sample was not clustered.}
\item{\code{clusterInfo}}{if sequential=TRUE and clusters were successfully found, a matrix of information regarding the algorithm behavior for each cluster (the starting and stopping K for each cluster, and the number of iterations for each cluster).}
\item{\code{whyStop}}{if sequential=TRUE and clusters were successfully found, a character string explaining what triggered the algorithm to stop.}
}
}
\description{
Given a nxp data matrix, this function will find clusters
}
\details{
If sequential=TRUE, the sequential clustering controls the 'k' argument of the underlying clustering so setting 'k=' in the list given to clusterDArgs or subsampleArgs will not do anything and will produce a warning to that effect.
}
\examples{
data(simData)
\dontrun{
#following code takes some time. 
#use clusterAll to do sequential clustering 
#(same as example in seqCluster only using clusterAll ...)
set.seed(44261)
clustSeqHier_v2<-clusterAll(simData,clusterFunction="hierarchical",
sequential=TRUE,subsample=TRUE,
subsampleArgs=list(resamp.n=100,samp.p=0.7,clusterFunction="kmeans",
clusterArgs=list(nstart=10)), seqArgs=list(beta=0.8,k0=5),
clusterDArgs=list(minSize=5))
}
#use clusterAll to do just clustering k=3 with no subsampling
clustNothing<-clusterAll(simData,clusterFunction="pam",subsample=FALSE,
sequential=FALSE, clusterDArgs=list(k=3))
}

