% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ClusterCells-class}
\alias{ClusterCells}
\alias{ClusterCells-class}
\alias{clusterCells}
\title{ClusterCells object and constructor}
\usage{
clusterCells(se, labels, isLog)
}
\arguments{
\item{se}{a matrix or \code{SummarizedExperiment} containing the clustered
data.}

\item{labels}{a vector with cluster labels.}
}
\value{
A \code{ClusterCells} object.
}
\description{
\code{ClusterCells} is a class that extends \code{SummarizedExperiment},
used to store the single-cell RNA-seq data and clustering information.
In addition to the slots of the \code{SummarizedExperiment} class, the
\code{ClusterCells} object has the following additional slots:
\itemize{
\item isLog: logical. Whether the data are in the linear or log scale.
\item labels: matrix. A matrix of cluster labels, useful for consensus
clustering.
\item dendrogram: dendrogram. A dendrogram containing the cluster
relationship.
\item coClustering: matrix. A matrix with the cluster co-occurrence
information; this can either be based on subsampling or on co-clustering
across parameter sets (see \code{compareChoices}).
}
}
\details{
The constructor \code{clusterCells} creates an object of the class
\code{ClusterCells}. However, the typical way of creating these objects is
the result of a call to \code{compareChoices} or \code{clusterAll}.

Note that when subsetting the data, the co-clustering and dendrogram
information are lost.
}
\examples{

se <- matrix(data=rnorm(200), ncol=10)
labels <- gl(5, 2)

cc <- clusterCells(se, as.numeric(labels), isLog = TRUE)

}

