% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ClusterExperiment-class}
\alias{ClusterExperiment}
\alias{ClusterExperiment-class}
\alias{clusterExperiment}
\alias{clusterExperiment,SummarizedExperiment,character-method}
\alias{clusterExperiment,SummarizedExperiment,factor-method}
\alias{clusterExperiment,SummarizedExperiment,matrix-method}
\alias{clusterExperiment,SummarizedExperiment,numeric-method}
\alias{clusterExperiment,matrix,ANY-method}
\title{Class ClusterExperiment}
\usage{
clusterExperiment(se, clusters, ...)

\S4method{clusterExperiment}{matrix,ANY}(se, clusters, ...)

\S4method{clusterExperiment}{SummarizedExperiment,numeric}(se, clusters, ...)

\S4method{clusterExperiment}{SummarizedExperiment,character}(se, clusters, ...)

\S4method{clusterExperiment}{SummarizedExperiment,factor}(se, clusters, ...)

\S4method{clusterExperiment}{SummarizedExperiment,matrix}(se, clusters,
  transformation, clusterType = "User", clusterInfo = NULL)
}
\arguments{
\item{se}{a matrix or \code{SummarizedExperiment} containing the data to be
clustered.}

\item{clusters}{can be either a numeric or character vector, a factor, or a
numeric matrix, containing the cluster labels.}

\item{...}{The arguments \code{transformation}, \code{clusterType} and
\code{clusterInfo} to be passed to the constructor for signature
\code{SummarizedExperiment,matrix}.}

\item{transformation}{function. A function to transform the data before
performing steps that assume normal-like data (i.e. constant variance), such
as the log.}

\item{clusterType}{a string describing the nature of the clustering. The
values `clusterSingle`, `clusterMany`, `mergeClusters`, `combineMany` are
reserved for the clustering coming from the package workflow and should not
be used when creating a new object with the constructor.}

\item{clusterInfo}{a list with information on the clustering (see Slots).}
}
\value{
A \code{ClusterExperiment} object.
}
\description{
\code{ClusterExperiment} is a class that extends
\code{SummarizedExperiment} and is used to store the data
and clustering information.

In addition to the slots of the \code{SummarizedExperiment}
class, the \code{ClusterExperiment} object has the additional slots described
in the Slots section.

There are several methods implemented for this class. The most
important methods (e.g., \code{\link{clusterMany}}, \code{\link{combineMany}},
...) have their own help page. Simple helper methods are described in the
Methods section. For a comprehensive list of methods specific to this class
see the Reference Manual.

The constructor \code{clusterExperiment} creates an object of
the class \code{ClusterExperiment}. However, the typical way of creating
these objects is the result of a call to \code{\link{clusterMany}} or
\code{\link{clusterSingle}}.

Note that when subsetting the data, the co-clustering and
dendrogram information are lost.
}
\section{Slots}{

\describe{
\item{\code{transformation}}{function. Function to transform the data by when methods
that assume normal-like data (e.g. log)}

\item{\code{clusterMatrix}}{matrix. A matrix giving the integer-valued cluster ids
for each sample. The rows of the matrix correspond to clusterings and columns
to samples. The integer values are assigned in the order that the clusters
were found, if found by setting sequential=TRUE in clusterSingle. "-1" indicates
the sample was not clustered.}

\item{\code{primaryIndex}}{numeric. An index that specifies the primary set of
labels.}

\item{\code{clusterInfo}}{list. A list with info about the clustering.
If created from \code{\link{clusterSingle}}, clusterInfo will include the
parameter used for the call, and the call itself. If \code{sequential = TRUE}
it will also include the following components.
\itemize{
\item{\code{clusterInfo}}{if sequential=TRUE and clusters were successfully
found, a matrix of information regarding the algorithm behavior for each
cluster (the starting and stopping K for each cluster, and the number of
iterations for each cluster).}
\item{\code{whyStop}}{if sequential=TRUE and clusters were successfully
found, a character string explaining what triggered the algorithm to stop.}
}}

\item{\code{clusterType}}{character vector with the origin of each column of
clusterMatrix.}

\item{\code{dendro_samples}}{dendrogram. A dendrogram containing the cluster
relationship (leaves are samples; see \code{\link{makeDendrogram}} for
details).}

\item{\code{dendro_clusters}}{dendrogram. A dendrogram containing the cluster
relationship (leaves are clusters; see \code{\link{makeDendrogram}} for
details).}

\item{\code{coClustering}}{matrix. A matrix with the cluster co-occurrence
information; this can either be based on subsampling or on co-clustering
across parameter sets (see \code{clusterMany}). The matrix is a square matrix
with number of rows/columns equal to the number of samples.}

\item{\code{clusterLegend}}{a list, one per cluster in \code{clusterMatrix}. Each
element of the list is a matrix with nrows equal to the number of different
clusters in the clustering, and consisting of at least two columns with the
following column names: "clusterId" and "color".}

\item{\code{orderSamples}}{a numeric vector (of integers) defining the order of
samples to be used for plotting of samples. Usually set internally by other
functions.}
}}
\examples{

se <- matrix(data=rnorm(200), ncol=10)
labels <- gl(5, 2)

cc <- clusterExperiment(se, as.numeric(labels), transformation =
function(x){x})

}

