% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelper.R
\docType{methods}
\name{ClusterExperiment-methods}
\alias{ClusterExperiment-methods}
\alias{[,ClusterExperiment,ANY,ANY,ANY-method}
\alias{[,ClusterExperiment,ANY,character,ANY-method}
\alias{[,ClusterExperiment,ANY,logical,ANY-method}
\alias{[,ClusterExperiment,ANY,numeric,ANY-method}
\alias{show,ClusterExperiment-method}
\alias{transformation,ClusterExperiment-method}
\alias{transformation}
\alias{transformation<-,ClusterExperiment,function-method}
\alias{transformation<-}
\alias{nClusterings,ClusterExperiment-method}
\alias{nClusterings}
\alias{nClusters,ClusterExperiment-method}
\alias{nClusters}
\alias{nFeatures,ClusterExperiment-method}
\alias{nFeatures}
\alias{nSamples,ClusterExperiment-method}
\alias{nSamples}
\alias{clusterMatrixNamed,ClusterExperiment-method}
\alias{clusterMatrixNamed}
\alias{clusterMatrixColors,ClusterExperiment-method}
\alias{clusterMatrixColors}
\alias{clusterMatrix,ClusterExperiment,missing-method}
\alias{clusterMatrix}
\alias{clusterMatrix,ClusterExperiment,numeric-method}
\alias{clusterMatrix}
\alias{clusterMatrix,ClusterExperiment,character-method}
\alias{clusterMatrix}
\alias{primaryCluster,ClusterExperiment-method}
\alias{primaryCluster}
\alias{primaryClusterIndex,ClusterExperiment-method}
\alias{primaryClusterIndex}
\alias{primaryClusterLabel,ClusterExperiment-method}
\alias{primaryClusterLabel}
\alias{primaryClusterNamed,ClusterExperiment-method}
\alias{primaryClusterNamed}
\alias{primaryClusterType,ClusterExperiment-method}
\alias{primaryClusterType}
\alias{subsetByCluster,ClusterExperiment-method}
\alias{subsetByCluster}
\alias{primaryClusterIndex<-,ClusterExperiment,numeric-method}
\alias{primaryClusterIndex<-}
\alias{dendroClusterIndex,ClusterExperiment-method}
\alias{dendroClusterIndex}
\alias{coClustering,ClusterExperiment-method}
\alias{coClustering}
\alias{coClustering<-,ClusterExperiment,matrix-method}
\alias{coClustering<-}
\alias{clusterTypes,ClusterExperiment-method}
\alias{clusterTypes}
\alias{clusteringInfo,ClusterExperiment-method}
\alias{clusteringInfo}
\alias{clusterLabels,ClusterExperiment-method}
\alias{clusterLabels}
\alias{clusterLabels<-,ClusterExperiment,character-method}
\alias{clusterLabels<-}
\alias{clusterLegend,ClusterExperiment-method}
\alias{clusterLegend}
\alias{clusterLegend<-,ClusterExperiment,list-method}
\alias{clusterLegend<-}
\alias{renameClusters,ClusterExperiment,character-method}
\alias{renameClusters}
\alias{recolorClusters,ClusterExperiment,character-method}
\alias{recolorClusters}
\alias{orderSamples,ClusterExperiment-method}
\alias{orderSamples}
\alias{orderSamples<-,ClusterExperiment,numeric-method}
\alias{orderSamples<-}
\alias{clusterTypes<-,ClusterExperiment,character-method}
\alias{clusterTypes<-}
\title{Helper methods for the ClusterExperiment class}
\usage{
\S4method{[}{ClusterExperiment,ANY,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ClusterExperiment,ANY,logical,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ClusterExperiment,ANY,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{ClusterExperiment}(object)

\S4method{transformation}{ClusterExperiment}(x)

\S4method{transformation}{ClusterExperiment,`function`}(object) <- value

\S4method{nClusterings}{ClusterExperiment}(x)

\S4method{nClusters}{ClusterExperiment}(x, ignoreUnassigned = TRUE)

\S4method{nFeatures}{ClusterExperiment}(x)

\S4method{nSamples}{ClusterExperiment}(x)

\S4method{clusterMatrixNamed}{ClusterExperiment}(x, whichClusters = "all")

\S4method{clusterMatrixColors}{ClusterExperiment}(x, whichClusters)

\S4method{clusterMatrix}{ClusterExperiment,missing}(x, whichClusters)

\S4method{clusterMatrix}{ClusterExperiment,numeric}(x, whichClusters)

\S4method{clusterMatrix}{ClusterExperiment,character}(x, whichClusters)

\S4method{primaryCluster}{ClusterExperiment}(x)

\S4method{primaryClusterIndex}{ClusterExperiment}(x)

\S4method{primaryClusterLabel}{ClusterExperiment}(x)

\S4method{primaryClusterNamed}{ClusterExperiment}(x)

\S4method{primaryClusterType}{ClusterExperiment}(x)

\S4method{subsetByCluster}{ClusterExperiment}(x, clusterValue,
  whichCluster = "primary", matchTo = c("name", "clusterIds"))

\S4method{primaryClusterIndex}{ClusterExperiment,numeric}(object) <- value

\S4method{dendroClusterIndex}{ClusterExperiment}(x)

\S4method{coClustering}{ClusterExperiment}(x)

\S4method{coClustering}{ClusterExperiment,matrix}(object) <- value

\S4method{clusterTypes}{ClusterExperiment}(x)

\S4method{clusteringInfo}{ClusterExperiment}(x)

\S4method{clusterLabels}{ClusterExperiment}(x)

\S4method{clusterLabels}{ClusterExperiment,character}(object) <- value

\S4method{clusterLegend}{ClusterExperiment}(x)

\S4method{clusterLegend}{ClusterExperiment,list}(object) <- value

\S4method{renameClusters}{ClusterExperiment,character}(object, value,
  whichCluster = "primary")

\S4method{recolorClusters}{ClusterExperiment,character}(object, value,
  whichCluster = "primary")

\S4method{orderSamples}{ClusterExperiment}(x)

\S4method{orderSamples}{ClusterExperiment,numeric}(object) <- value

\S4method{clusterTypes}{ClusterExperiment,character}(object) <- value
}
\arguments{
\item{x, object}{a ClusterExperiment object.}

\item{..., i, j, drop}{Forwarded to the
\code{\link{SingleCellExperiment}} method.}

\item{value}{The value to be substituted in the corresponding slot. See the
slot descriptions in \code{\link{ClusterExperiment}} for details on what
objects may be passed to these functions.}

\item{ignoreUnassigned}{logical. If true, ignore the clusters with -1 or -2 assignments in calculating the number of clusters per clustering.}

\item{whichClusters}{argument that can be either numeric or
character value. If numeric, gives the indices of the \code{clusterMatrix}
to return; this can also be used to defined an ordering for the
clusterings. \code{whichClusters} can be a character value identifying the 
\code{clusterTypes} to be used, or if not matching \code{clusterTypes} then
\code{clusterLabels}; alternatively \code{whichClusters} can be either 
'all' or 'workflow' to indicate choosing all clusters or choosing all 
\code{\link{workflowClusters}}. If missing, the entire matrix of all
clusterings is returned.}

\item{clusterValue}{values of the cluster to match to for subsetting}

\item{whichCluster}{argument to identify cluster, taking input like 
\code{whichClusters}, only requires that only a single clustering can be identified.}

\item{matchTo}{for subsetting, whether to match to the cluster name
(\code{"name"}) or internal cluster id (\code{"clusterIds"})}
}
\value{
\code{transformation} prints the function used to transform the data
prior to clustering.

\code{nClusterings} returns the number of clusterings (i.e., ncol of
clusterMatrix).

\code{nClusters} returns the number of clusters per clustering

\code{nFeatures} returns the number of features (same as `nrow`).

\code{nSamples} returns the number of samples (same as `ncol`).

\code{clusterMatrixNamed} returns a matrix with cluster labels.

\code{clusterMatrixColors} returns the matrix with all the clusterings, using the internally stored colors for each cluster

\code{clusterMatrix} returns the matrix with all the clusterings.

\code{clusterMatrix} returns the matrix with all the clusterings.

\code{clusterMatrix} returns the matrix with all the clusterings.

\code{primaryCluster} returns the primary clustering (as numeric).

\code{primaryClusterIndex} returns/sets the primary clustering index
(i.e., which column of clusterMatrix corresponds to the primary clustering).

\code{primaryClusterIndex} returns/sets the primary clustering index
(i.e., which column of clusterMatrix corresponds to the primary clustering).

\code{primaryClusterNamed} returns the primary cluster (using cluster
labels).

\code{primaryClusterIndex} returns/sets the primary clustering index
(i.e., which column of clusterMatrix corresponds to the primary clustering).

\code{subsetByCluster} subsets the object by clusters in a clustering
and returns a ClusterExperiment object with only those samples

\code{dendroClusterIndex} returns/sets the clustering index 
of the clusters used to create dendrogram
(i.e., which column of clusterMatrix corresponds to the clustering).

\code{coClustering} returns/sets the co-clustering matrix.

\code{clusterTypes} returns/sets the clusterTypes slot.

\code{clusteringInfo} returns the clusterInfo slot.

\code{clusterLabels} returns/sets the column names of the clusterMatrix slot.

\code{clusterLegend} returns/sets the clusterLegend slot.

\code{renameClusters} changes the names assigned to clusters within a clustering

\code{recolorClusters} changes the colors assigned to clusters within a clustering

\code{orderSamples} returns/sets the orderSamples slot.
}
\description{
This is a collection of helper methods for the ClusterExperiment class.
}
\details{
Note that when subsetting the data, the dendrogram information and
the co-clustering matrix are lost.

Note that redefining the transformation function via
  \code{transformation(x)<-} will check the validity of the transformation on
  the data assay. If the assay is large, this may be time consuming. Consider
  using a call to ClusterExperiment, which has the option as to whether to
  check the validity of the transformation.
}
