% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelper.R
\docType{methods}
\name{ClusterExperiment-methods}
\alias{ClusterExperiment-methods}
\alias{[,ClusterExperiment,ANY,ANY,ANY-method}
\alias{[,ClusterExperiment,ANY,logical,ANY-method}
\alias{[,ClusterExperiment,ANY,numeric,ANY-method}
\alias{show,ClusterExperiment-method}
\alias{clusterMatrixNamed,ClusterExperiment-method}
\alias{clusterMatrixNamed}
\alias{primaryClusterNamed,ClusterExperiment-method}
\alias{primaryClusterNamed}
\alias{transformation,ClusterExperiment-method}
\alias{transformation}
\alias{nClusters,ClusterExperiment-method}
\alias{nClusters}
\alias{nFeatures,ClusterExperiment-method}
\alias{nSamples,ClusterExperiment-method}
\alias{clusterMatrix,ClusterExperiment-method}
\alias{clusterMatrix}
\alias{primaryCluster,ClusterExperiment-method}
\alias{primaryCluster}
\alias{primaryClusterIndex,ClusterExperiment-method}
\alias{primaryClusterIndex}
\alias{primaryClusterIndex<-,ClusterExperiment,numeric-method}
\alias{primaryClusterIndex<-}
\alias{coClustering,ClusterExperiment-method}
\alias{coClustering}
\alias{coClustering<-,ClusterExperiment,matrix-method}
\alias{coClustering<-}
\alias{clusterTypes,ClusterExperiment-method}
\alias{clusterTypes}
\alias{clusterInfo,ClusterExperiment-method}
\alias{clusterInfo}
\alias{clusterLabels,ClusterExperiment-method}
\alias{clusterLabels}
\alias{clusterLabels<-,ClusterExperiment,character-method}
\alias{clusterLabels<-}
\alias{clusterLegend,ClusterExperiment-method}
\alias{clusterLegend}
\alias{clusterLegend<-,ClusterExperiment,list-method}
\alias{clusterLegend<-}
\alias{orderSamples,ClusterExperiment-method}
\alias{orderSamples}
\alias{orderSamples<-,ClusterExperiment,numeric-method}
\alias{orderSamples<-}
\alias{clusterTypes<-,ClusterExperiment,character-method}
\alias{clusterTypes<-}
\title{Helper methods for the ClusterExperiment class}
\usage{
\S4method{[}{ClusterExperiment,ANY,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ClusterExperiment,ANY,logical,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ClusterExperiment,ANY,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{ClusterExperiment}(object)

\S4method{clusterMatrixNamed}{ClusterExperiment}(x)

\S4method{primaryClusterNamed}{ClusterExperiment}(x)

\S4method{transformation}{ClusterExperiment}(x)

\S4method{nClusters}{ClusterExperiment}(x)

\S4method{nFeatures}{ClusterExperiment}(x)

\S4method{nSamples}{ClusterExperiment}(x)

\S4method{clusterMatrix}{ClusterExperiment}(x)

\S4method{primaryCluster}{ClusterExperiment}(x)

\S4method{primaryClusterIndex}{ClusterExperiment}(x)

\S4method{primaryClusterIndex}{ClusterExperiment,numeric}(object) <- value

\S4method{coClustering}{ClusterExperiment}(x)

\S4method{coClustering}{ClusterExperiment,matrix}(object) <- value

\S4method{clusterTypes}{ClusterExperiment}(x)

\S4method{clusterInfo}{ClusterExperiment}(x)

\S4method{clusterLabels}{ClusterExperiment}(x)

\S4method{clusterLabels}{ClusterExperiment,character}(object) <- value

\S4method{clusterLegend}{ClusterExperiment}(x)

\S4method{clusterLegend}{ClusterExperiment,list}(object) <- value

\S4method{orderSamples}{ClusterExperiment}(x)

\S4method{orderSamples}{ClusterExperiment,numeric}(object) <- value

\S4method{clusterTypes}{ClusterExperiment,character}(object) <- value
}
\arguments{
\item{x, object}{a ClusterExperiment object.}

\item{..., i, j, drop}{Forwarded to the
\code{\link[SummarizedExperiment]{SummarizedExperiment}} method.}

\item{value}{The value to be substituted in the corresponding slot. See the
slot descriptions in \code{\link{ClusterExperiment}} for details on what
objects may be passed to these functions.}
}
\value{
\code{clusterMatrixNamed} returns a matrix with cluster labels.

\code{primaryClusterNamed} returns the primary cluster (using cluster
labels).

\code{transformation} prints the function used to transform the data
prior to clustering.

\code{nClusters} returns the number of clusterings (i.e., ncol of
clusterMatrix).

\code{nFeatures} returns the number of features (same as `nrow`).

\code{nSamples} returns the number of samples (same as `ncol`).

\code{clusterMatrix} returns the matrix with all the clusterings.

\code{primaryCluster} returns the primary clustering (as numeric).

\code{primaryClusterIndex} returns/sets the primary clustering index
(i.e., which column of clusterMatrix corresponds to the primary clustering).

\code{coClustering} returns/sets the co-clustering matrix.

\code{clusterTypes} returns/sets the clusterTypes slot.

\code{clusterInfo} returns the clusterInfo slot.

\code{clusterLabels} returns/sets the column names of the clusterMatrix slot.

\code{clusterLegend} returns/sets the clusterLegend slot.

\code{orderSamples} returns/sets the orderSamples slot.
}
\description{
This is a collection of helper methods for the ClusterExperiment class.
}
\details{
Note that when subsetting the data, the dendrogram information and
the co-clustering matrix are lost.
}
