% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperClusterFunction.R
\docType{methods}
\name{ClusterFunction-methods}
\alias{ClusterFunction-methods}
\alias{requiredArgs,ClusterFunction-method}
\alias{requiredArgs,character-method}
\alias{requiredArgs}
\alias{requiredArgs,character-method}
\alias{requiredArgs,factor-method}
\alias{algorithmType,ClusterFunction-method}
\alias{algorithmType}
\alias{algorithmType,character-method}
\alias{algorithmType,factor-method}
\alias{inputType,ClusterFunction-method}
\alias{inputType}
\alias{inputType,character-method}
\alias{inputType,factor-method}
\title{Helper methods for the ClusterFunction class}
\usage{
\S4method{requiredArgs}{character}(object)

\S4method{requiredArgs}{ClusterFunction}(object, genericOnly = FALSE)

\S4method{requiredArgs}{character}(object)

\S4method{requiredArgs}{character}(object)

\S4method{requiredArgs}{factor}(object)

\S4method{algorithmType}{ClusterFunction}(object)

\S4method{algorithmType}{character}(object)

\S4method{algorithmType}{factor}(object)

\S4method{inputType}{ClusterFunction}(object)

\S4method{inputType}{character}(object)

\S4method{inputType}{factor}(object)
}
\arguments{
\item{genericOnly}{logical If TRUE, return only the generic required arguments (i.e. those required by the algorithm type) and not the arguments specific to that clustering found in the slot \code{requiredArgs}. If FALSE both sets of arguments are returned.}
}
\description{
This is a collection of helper methods for the ClusterExperiment class.
}
\details{
Note that when subsetting the data, the dendrogram information and
the co-clustering matrix are lost.
}
