% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperClusterFunction.R
\docType{methods}
\name{ClusterFunction-methods}
\alias{ClusterFunction-methods}
\alias{requiredArgs,ClusterFunction-method}
\alias{requiredArgs,list-method}
\alias{requiredArgs,character-method}
\alias{requiredArgs}
\alias{requiredArgs,character-method}
\alias{requiredArgs,factor-method}
\alias{algorithmType,ClusterFunction-method}
\alias{algorithmType}
\alias{algorithmType,character-method}
\alias{algorithmType,factor-method}
\alias{algorithmType,list-method}
\alias{inputType,ClusterFunction-method}
\alias{inputType}
\alias{inputType,list-method}
\alias{inputType,character-method}
\alias{inputType,factor-method}
\title{Helper methods for the ClusterFunction class}
\usage{
\S4method{requiredArgs}{character}(object)

\S4method{requiredArgs}{ClusterFunction}(object, genericOnly = FALSE)

\S4method{requiredArgs}{list}(object, ...)

\S4method{requiredArgs}{character}(object)

\S4method{requiredArgs}{character}(object)

\S4method{requiredArgs}{factor}(object)

\S4method{algorithmType}{ClusterFunction}(object)

\S4method{algorithmType}{character}(object)

\S4method{algorithmType}{factor}(object)

\S4method{algorithmType}{list}(object, ...)

\S4method{inputType}{ClusterFunction}(object)

\S4method{inputType}{list}(object, ...)

\S4method{inputType}{character}(object)

\S4method{inputType}{factor}(object)
}
\arguments{
\item{object}{input to the method, either a \code{ClusterFunction}
class or a character describing a built-in \code{ClusterFunction} object. 
Can also be a \code{list} of \code{ClusterFunction} objects, in which case 
the list must have names for each function.}

\item{genericOnly}{logical If TRUE, return only the generic required
arguments (i.e. those required by the algorithm type) and not the arguments
specific to that clustering found in the slot \code{requiredArgs}. If FALSE
both sets of arguments are returned.}
}
\value{
\code{requiredArgs}  returns a list of the required args of a
  function (via a call to \code{\link{requiredArgs}})

\code{algorithmType} returns a character value giving the type of 
clustering function ("01" or "K")

\code{inputType} returns a character value giving the input 
type of the object
}
\description{
This is a collection of helper methods for the ClusterExperiment class.
}
\details{
Note that when subsetting the data, the dendrogram information and 
  the co-clustering matrix are lost.
}
