% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusterHclust.R
\name{clusterHclust}
\alias{clusterHclust}
\title{Make hierarchy of set of clusters}
\usage{
clusterHclust(dat, cl, full = TRUE, unassigned = c("remove", "outgroup",
  "cluster"), ...)
}
\arguments{
\item{dat}{data to define the mediods from. Samples on rows, features on columns}

\item{cl}{A numeric vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.}

\item{full}{Logical as to whether to expand to include all samples as tips (useful for heatmap), or to have dendrogram just of the clusters (useful for getBestGenes)}

\item{unassigned}{what to do with with unassigned ("-1") cluster; only relevant if full=TRUE. Default is to remove those samples, so that only get heatmap of clustered observations. ``outgroup'' puts the unassigned into an outgroup on the dendrogram. ``cluster'' clusters them with the mediods.}

\item{...}{arguments passed to hclust}
}
\value{
A dendrogram for the clusters. If full=TRUE, the number of leaves is equal to the number of samples. If full=FALSE, the number of leaves is equal to the number of clusters.
}
\description{
Makes a dendrogram of a set of clusters based on hclust on the mediods of the cluster.
}
\details{
If full=TRUE, then the expanded dendrogram is created by giving all of the samples the mediod of the cluster and applying hclust to it; if further unassigned=="cluster", then the expanded dendrogram is created by hclust of the expanded mediod data plus the original unclustered observations.
}
\examples{
data(simData)
#create a clustering, for 8 clusters (truth was 3)
cl<-clusterAll(simData,clusterFunction="pam",subsample=FALSE,
sequential=FALSE, clusterDArgs=list(k=8))$cl
#create dendrogram of clusters:
hcl<-clusterHclust(dat=simData,cl,full=FALSE)
plot(hcl)

#create dendrogram for plotting with data in heatmap:
hclData<-clusterHclust(dat=simData,cl,full=TRUE)
dualHeatmap(cl,heatData=simCount,clusterData=hclData,colorScale=seqPal5,
	annCol=data.frame(PAM8=cl,Truth=trueCluster))
}

