% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelper.R
\docType{methods}
\name{clusterLabels<-,ClusterExperiment,character-method}
\alias{clusterLabels,ClusterExperiment,character-method}
\alias{clusterLabels,ClusterExperiment,missing-method}
\alias{clusterLabels,ClusterExperiment,numeric-method}
\alias{clusterLabels<-,ClusterExperiment,character-method}
\title{Methods to set and return the clustering labels}
\usage{
\S4method{clusterLabels}{ClusterExperiment,character}(object) <- value

\S4method{clusterLabels}{ClusterExperiment,numeric}(x, whichClusters)

\S4method{clusterLabels}{ClusterExperiment,character}(x,
  whichClusters = "all")

\S4method{clusterLabels}{ClusterExperiment,missing}(x, whichClusters)
}
\arguments{
\item{`whichClusters`}{controls which labels to be returned. It is either
numeric, in which case gives the indices of the clusters, or character, in
which case it matches to clusterType(x) to find the indices of the clusters.}
}
\description{
These methods will return or set the clustering labels (i.e., the
colnames of the clusterMatrix slot).
}

