% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterLabels.R
\docType{methods}
\name{clusterLabels,ClusterExperiment,numeric-method}
\alias{clusterLabels}
\alias{clusterLabels,ClusterExperiment,character-method}
\alias{clusterLabels,ClusterExperiment,missing-method}
\alias{clusterLabels,ClusterExperiment,numeric-method}
\alias{clusterLabels<-}
\alias{clusterLabels<-,ClusterExperiment,character-method}
\title{Methods to set and return the clustering labels}
\usage{
\S4method{clusterLabels}{ClusterExperiment,numeric}(x, whichClusters)

\S4method{clusterLabels}{ClusterExperiment,character}(object) <- value

\S4method{clusterLabels}{ClusterExperiment,character}(x,
  whichClusters = "all")

\S4method{clusterLabels}{ClusterExperiment,missing}(x, whichClusters)
}
\arguments{
\item{x, object}{a \code{\link{ClusterExperiment}} object.}

\item{whichClusters}{controls which labels to be returned. It is either
numeric, in which case gives the indices of the clusters, or character, in
which case it matches to \code{clusterType(x)} to find the indices of the
clusters.}

\item{value}{character. A vector of length equal to
\code{NCOL(clusterMatrix(object))}.}
}
\value{
A vector with the clustering labels.
}
\description{
These methods will return or set the clustering labels (i.e., the
colnames of the clusterMatrix slot).
}
\examples{
data(simData)

cl <- clusterMany(simData,nPCADims=c(5,10,50),  dimReduce="PCA",
clusterFunction="pam", ks=2:4, findBestK=c(FALSE), removeSil=TRUE,
subsample=FALSE)

clusterLabels(cl)
clusterLabels(cl, whichClusters=1:3)

clusterType(cl)
clusterLabels(cl, whichClusters="clusterMany")
}

