% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSingle.R
\docType{methods}
\name{clusterSingle}
\alias{clusterSingle}
\alias{clusterSingle,ClusterExperiment-method}
\alias{clusterSingle,SummarizedExperiment-method}
\alias{clusterSingle,matrix-method}
\alias{clusterSingle-methods}
\title{General wrapper method to cluster the data}
\usage{
\S4method{clusterSingle}{matrix}(x, subsample = TRUE, sequential = FALSE,
  clusterFunction = c("tight", "hierarchical01", "pam", "hierarchicalK"),
  clusterDArgs = NULL, subsampleArgs = NULL, seqArgs = NULL,
  isCount = FALSE, transFun = NULL, dimReduce = c("none", "PCA", "var"),
  ndims = NA)

\S4method{clusterSingle}{SummarizedExperiment}(x, ...)

\S4method{clusterSingle}{ClusterExperiment}(x, ...)
}
\arguments{
\item{x}{the data on which to run the clustering (features in rows).}

\item{subsample}{logical as to whether to subsample via 
\code{\link{subsampleClustering}} to get the distance matrix at each 
iteration; otherwise the distance function will be determined by argument
\code{distFunction} passed in \code{clusterDArgs}.}

\item{sequential}{logical whether to use the sequential strategy (see
Details).}

\item{clusterFunction}{passed to \code{\link{clusterD}} option
'clusterFunction' to indicate method of clustering, see
\code{\link{clusterD}}.}

\item{clusterDArgs}{list of additional arguments to be passed to
\code{\link{clusterD}}.}

\item{subsampleArgs}{list of arguments to be passed to
\code{\link{subsampleClustering}}.}

\item{seqArgs}{list of additional arguments to be passed to
\code{\link{seqCluster}}.}

\item{isCount}{logical. Whether the data are in counts, in which case the
default \code{transFun} argument is set as log2(x+1). This is simply a
convenience to the user, and can be overridden by giving an explicit
function to \code{transFun}.}

\item{transFun}{function A function to use to transform the input data matrix
before clustering.}

\item{dimReduce}{character A character identifying what type of 
dimensionality reduction to perform before clustering. Options are 
"none","PCA", "var","cv", and "mad". See \code{\link{transform}} for more
details.}

\item{ndims}{integer An integer identifying how many dimensions to reduce to
in the reduction specified by \code{dimReduce}}

\item{...}{arguments to be passed on to the method for signature
\code{matrix}.}
}
\value{
A \code{\link{ClusterExperiment}} object.
}
\description{
Given a data matrix, \code{\link{SummarizedExperiment}}, or
\code{\link{ClusterExperiment}} object, this function will find clusters,
based on a single specification of parameters.
}
\details{
If sequential=TRUE, the sequential clustering controls the 'k'
  argument of the underlying clustering so setting 'k=' in the list given to
  clusterDArgs or subsampleArgs will not do anything and will produce a
  warning to that effect.
}
\examples{
data(simData)

\dontrun{
#following code takes some time.
#use clusterSingle to do sequential clustering
#(same as example in seqCluster only using clusterSingle ...)
set.seed(44261)
clustSeqHier_v2 <- clusterSingle(simData, clusterFunction="hierarchical01",
sequential=TRUE, subsample=TRUE, subsampleArgs=list(resamp.n=100, samp.p=0.7,
clusterFunction="kmeans", clusterArgs=list(nstart=10)),
seqArgs=list(beta=0.8, k0=5), clusterDArgs=list(minSize=5))
}

#use clusterSingle to do just clustering k=3 with no subsampling
clustNothing <- clusterSingle(simData, clusterFunction="pam",
subsample=FALSE, sequential=FALSE, clusterDArgs=list(k=3))
}
\seealso{
\code{\link{clusterMany}} to compare multiple choices of parameters.
}

