% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSharedClusters.R
\name{findSharedClusters}
\alias{findSharedClusters}
\title{Find sets of samples that stay together across clusterings
Find sets of samples that stay together across clusterings in order to define a new clustering vector.}
\usage{
findSharedClusters(clusterMat, proportion = 1,
  clusterFunction = "hierarchical", propUnassigned = 0.5, plot = FALSE,
  minSize = 5)
}
\arguments{
\item{clusterMat}{A matrix with samples on rows and different clustering assignments on each column.}

\item{proportion}{The proportion of times that two sets of samples should be together in order to be grouped into a cluster (if <1, passed to clusterD via alpha=1-proportion)}

\item{clusterFunction}{the clustering to use (passed to \code{\link{clusterD}}); currently must be of type '01'.}

\item{propUnassigned}{samples with greater than this proportion of assignments equal to '-1' are assigned a '-1' cluster value.}

\item{plot}{logical. If true, will create a heatmap of the shared percentages with the final cluster found by function}

\item{minSize}{minimum size required for a set of samples to be considered in a cluster because of shared clustering, passed to \code{\link{clusterD}}}
}
\value{
A list with values 
\itemize{

\item{\code{clustering}}{vector of cluster assignments, with "-1" implying unassigned}

\item{\code{percentageShared}}{a nsample x nsample matrix of the percent co-occurance across clusters used to find the final clusters. Percentage is out of those not '-1'}
\item{\code{noUnassignedCorrection}{a vector of cluster assignments before samples were converted to '-1' because had >\code{propUnassigned} '-1' values (i.e. the direct output of the \code{clusterD} output.)}}
}
}
\description{
Find sets of samples that stay together across clusterings
Find sets of samples that stay together across clusterings in order to define a new clustering vector.
}
\details{
The function tries to find a consensus cluster across many different clusterings of the same sample. It does so by creating a nsample x nsample matrix of the percentage of co-occurance of each sample and then calling clusterD to cluster the co-occurance matrix. The function assumes that '-1' labels indicate clusters that are not assigned to a cluster. Co-occurance with the unassigned cluster is treated differently than other clusters. The percent co-occurance is taken only with respect to those clusterings where both samples were assigned. Then samples with more than \code{propUnassigned} values that are '-1' across all of the clusterings are assigned a '-1' regardless of their cluster assignment. 

If the majority
}

