% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getBestGenes}
\alias{getBestGenes}
\title{Function for finding best genes associated with clusters}
\usage{
getBestGenes(cl, dat, type = c("F", "Dendro", "Pairs"), dendro = NULL,
  pairMat = NULL, returnType = c("Index", "Table"),
  contrastAdj = c("PerContrast", "AfterF", "All"), ...)
}
\arguments{
\item{cl}{A vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.}

\item{dat}{data for the test, samples in rows}

\item{type}{What type of test to do. `F' gives the omnibus F-statistic, `Dendro' traverses the given dendrogram and does contrasts of the samples in each side, and `Pairs' does pair-wise contrasts based on the pairs given in pairMat.}

\item{dendro}{The dendrogram to traverse if type="Dendro". Note that this should be the dendrogram of the clusters, not of the individual samples.}

\item{pairMat}{matrix giving the pairs of clusters for which to do pair-wise contrasts (must match to elements of cl). If NULL, will do all pairwise of the clusters in \code{cl} (excluding "-1" categories). Each row is a pair to be compared and must match the names of the clusters in the vector \code{cl}.}

\item{returnType}{Whether to return the index of genes, or the full table given by topTable or topTableF.}

\item{contrastAdj}{What type of FDR correction to do for contrasts tests (i.e. type='Dendro' or 'Pairs').}

\item{...}{options to pass to \code{\link{topTable}} or \code{\link{topTableF}} (see \code{\link{limma}} package)}
}
\value{
A data.frame in the same format as \code{\link{topTable}}, except that the column name \code{ProbeID} is changed to \code{Gene} and a column \code{IndexInOriginal} is provided to give the row index of the gene in the original dataset.
}
\description{
Calls limma on input data to determine genes most associated with found clusters (either pairwise comparisons, or following a tree that clusters the clusters).
}
\details{
getBestGenes returns best genes corresponding to a cluster assignment. It uses limma to fit the models, and limma's functions topTable or topTableF to find the best genes.  See the options of these functions to put better control on what gets returned (e.g. only if significant, only if log-fc is above a certain amount, etc.). In particular, set 'number=' to define how many significant genes to return (where number is per contrast for the 'Pairs' or 'Dendro' option)
}

