% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/makeDendrogram.R
\docType{methods}
\name{makeDendrogram}
\alias{makeDendrogram}
\alias{makeDendrogram,ClusterExperiment-method}
\alias{makeDendrogram,matrix-method}
\alias{plotDendrogram,ClusterExperiment-method}
\title{Make hierarchy of set of clusters}
\usage{
makeDendrogram(x, ...)

\S4method{makeDendrogram}{ClusterExperiment}(x, dimReduce = c("none", "PCA",
  "mostVar"), ndims = NA, ...)

\S4method{makeDendrogram}{matrix}(x, cluster,
  unassignedSamples = c("outgroup", "cluster", "remove"), ...)

\S4method{plotDendrogram}{ClusterExperiment}(x, leaves = c("samples",
  "clusters"), main, ...)
}
\arguments{
\item{x}{data to define the mediods from. Matrix and
\code{\link{ClusterExperiment}} supported.}

\item{...}{arguments passed to hclust}

\item{dimReduce}{character A character identifying what type of
dimensionality reduction to perform before clustering}

\item{ndims}{integer An integer identifying how many dimensions to reduce to
in the reduction specified by \code{dimReduce}}

\item{cluster}{A numeric vector with cluster assignments. If x is a
ClusterExperiment object, cluster is automatically the primaryCluster(x).
``-1'' indicates the sample was not assigned to a cluster.}

\item{unassigned}{how to handle unassigned samples("-1") ; only relevant
for sample clustering. See details.
#' @param isCount logical. Whether the data are in counts, in which case the
default \code{transFun} argument is set as log(x+1). This is simply a
convenience to the user, and can be overridden by giving an explicit function
to \code{transFun}.}

\item{transFun}{function A function to use to transform the input data matrix
before clustering.}
}
\value{
If x is a matrix, a list with two dendrograms, one in which the
leaves are clusters and one in which the leaves are samples. If x is a
ClusterExperiment object, the dendrograms are saved in the appropriate slots.
}
\description{
Makes a dendrogram of a set of clusters based on hclust on the
mediods of the cluster.
}
\details{
The function returns two dendrograms (as a list if x is a matrix) or
in the appropriate slots if x is ClusterExperiment. The cluster dendrogram
is created by applying \code{\link{hclust}} to the mediods of each cluster.
In the sample dendrogram the clusters are again clustered, but now the
samples are also part of the resulting dendrogram. This is done by giving
each sample the value of the mediod of its cluster.

The argument \code{unassignedSamples} governs what is done with
unassigned samples (defined by a -1 cluster value). If unassigned=="cluster",
then the dendrogram is created by hclust of the expanded mediod data plus the
original unclustered observations. If \code{unassignedSamples} is "outgroup",
then all unassigned samples are put as an outgroup.
}
\examples{
data(simData)
#create a clustering, for 8 clusters (truth was 3)
cl<-clusterSingle(simData,clusterFunction="pam",subsample=FALSE,
sequential=FALSE, clusterDArgs=list(k=8))$cl
#create dendrogram of clusters:
hcl<-clusterHclust(dat=simData,cl,full=FALSE)
plot(hcl)

#create dendrogram for plotting with data in heatmap:
hclData<-clusterHclust(dat=simData,cl,full=TRUE)
plotHeatmap(cl,heatData=simCount,clusterData=hclData,colorScale=seqPal5,
annCol=data.frame(PAM8=cl,Truth=trueCluster))
}

