% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFunction.R
\docType{methods}
\name{makeDimReduce}
\alias{makeDimReduce}
\alias{makeDimReduce,SingleCellExperiment-method}
\alias{makeDimReduce,matrix-method}
\alias{makeDimReduce,SummarizedExperiment-method}
\alias{makeDimReduce,ClusterExperiment-method}
\alias{listBuiltInReduceDims}
\title{Calculate dimensionality reduction of data}
\usage{
\S4method{makeDimReduce}{SingleCellExperiment}(object, reducedDims = "PCA",
  maxDims = 500, transFun = NULL, isCount = FALSE)

\S4method{makeDimReduce}{matrix}(object, ...)

\S4method{makeDimReduce}{SummarizedExperiment}(object, ...)

\S4method{makeDimReduce}{ClusterExperiment}(object, ...)

listBuiltInReduceDims()
}
\arguments{
\item{object}{input to use for the data for dimensionality reduction. Can be matrix, SummarizedExperiment, SingleCellExperiment, or ClusterExperiment object}

\item{reducedDims}{a vector of character values indicating the methods of dimensionality reduction to be performed. Currently only "PCA" is implemented.}

\item{maxDims}{Numeric vector of integer giving the number of PC dimensions to calculate. 
\code{maxDims} can also take values between (0,1) to indicate keeping the
number of PCA dimensions necessary to account for that proportion of the
variance. \code{maxDims} should be of same length as \code{reducedDims}, indicating the number of dimensions to keep for each method (if \code{maxDims} is of length 1, the same number of dimensions will be used for each).}

\item{transFun}{a transformation function to be applied to the data. If the transformation 
applied to the data creates an error or NA values, then the function will throw an error.
If object is of class \code{ClusterExperiment}, the stored transformation will be used 
and giving this parameter will result in an error.}

\item{isCount}{if \code{transFun=NULL}, then \code{isCount=TRUE} will determine the
will give a transformation function \code{function(x){x}}. Ignored if \code{transFun=NULL}.
If object is of class \code{ClusterExperiment}, the stored transformation will be used 
and giving this parameter will result in an error.}

\item{...}{Values passed on the the 'SingleCellExperiment' method.}
}
\value{
a SingleCellExperiment object with the indicated diminsionality reduction methods stored in the \code{reduceDims} slot.

A \code{\link{SingleCellExperiment}} containing the dimensionality reduction in the corresponding slots with names corresponding to the name given in \code{reducedDims}.
}
\description{
A Function for performing and storing common dimensionality reduction techniques
}
\details{
The PCA method uses either \code{prcomp} from the \code{stats} package or  \code{svds} from the \code{RSpectra} package to perform PCA. Both are called on \code{t(assay(x))} with \code{center=TRUE} and \code{scale=TRUE} (i.e. the feature are centered and scaled), so that
  it is performing PCA on the correlation matrix of the features.
}
\examples{
data(simData)
listBuiltInReduceDims()
scf<-makeDimReduce(simData, reducedDims="PCA", maxDims=3)
scf
}
