% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFunction.R
\docType{methods}
\name{makeFilterStats}
\alias{makeFilterStats}
\alias{makeFilterStats,matrix-method}
\alias{makeFilterStats,SummarizedExperiment-method}
\alias{makeFilterStats,SingleCellExperiment-method}
\alias{makeFilterStats,ClusterExperiment-method}
\alias{listBuiltInFilterStats}
\alias{filterData,SingleCellFilter-method}
\title{Calculate filtering statistics}
\usage{
\S4method{makeFilterStats}{SingleCellFilter}(object,
  filterStats = listBuiltInFilterStats(), transFun = NULL,
  isCount = FALSE)

\S4method{makeFilterStats}{matrix}(object, ...)

\S4method{makeFilterStats}{SummarizedExperiment}(object, ...)

\S4method{makeFilterStats}{SingleCellExperiment}(object, ...)

\S4method{makeFilterStats}{ClusterExperiment}(object,
  whichClusterIgnoreUnassigned = NULL,
  filterStats = listBuiltInFilterStats(), ...)

listBuiltInFilterStats()

\S4method{filterData}{SingleCellFilter}(object, type, cutoff, percentile,
  absolute = FALSE, keepLarge = TRUE)
}
\arguments{
\item{object}{object from which user wants to calculate per-row statistics}

\item{filterStats}{character vector of statistics to calculate. 
Must be one of the character values given by \code{listBuildInFilterStats()}.}

\item{transFun}{a transformation function to be applied to the data. If the transformation 
applied to the data creates an error or NA values, then the function will throw an error.
If object is of class \code{ClusterExperiment}, the stored transformation will be used 
and giving this parameter will result in an error.}

\item{isCount}{if \code{transFun=NULL}, then \code{isCount=TRUE} will determine the
will give a transformation function \code{function(x){x}}. Ignored if \code{transFun=NULL}.
If object is of class \code{ClusterExperiment}, the stored transformation will be used 
and giving this parameter will result in an error.}

\item{...}{Values passed on the the 'SingleCellFilter' method.}

\item{whichClusterIgnoreUnassigned}{indicates clustering that should be used to filter out unassigned samples from the calculations. If \code{NULL} no filtering of samples will be done. See details for more information.}

\item{cutoff}{numeric. A value at which to filter the rows (genes) for the test statistic}

\item{percentile}{numeric. Either a number between 0,1 indicating what percentage of the rows (genes) to keep or an integer value indicated the number of rows (genes) to keep}

\item{absolute}{whether to take the absolute value of the filter statistic}

\item{keepLarge}{logical whether to keep rows (genes) with large values of the test statistic or small values of the test statistic.}
}
\value{
A SingleCellFilter object with the rows (genes) removed based on filters
}
\description{
Function for calculating, per row (gene), built-in statistical 
   functions that might be used for filtering.
}
\details{
\code{whichClusterIgnoreUnassigned} is only an option when applied to a
\code{ClusterExperiment} classs and indicates that the filtering statistics should be
calculated based on samples that are unassigned by the
 designated clustering. The name given to the filter in this case is of the form
 \code{<filterStats>_<clusterLabel>}, i.e. the clustering label of the clustering is
 appended to the standard name for the filtering statistic.

Note that \code{filterData} returns a SingleCellFilter object. To get the actual data out use either assay or \code{\link{transformData}} if transformed data is desired.
}
\examples{
data(simData)
listBuiltInFilterStats()
scf<-makeFilterStats(simData,filterStats=c("var","mad"))
scf
scfFiltered<-filterData(scf,type="mad",percentile=10)
scfFiltered
assay(scfFiltered)[1:10,1:10]
}
