% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterHclust.R
\name{mergeClusters}
\alias{mergeClusters}
\title{Merge clusters based on dendrogram}
\usage{
mergeClusters(dat, cl, dendro = NULL, mergeMethod = c("none", "adjP",
  "locfdr", "MB", "JC"), cutoff = 0.1, plotType = c("none", "all",
  "mergeMethod"), countData = TRUE, ...)
}
\arguments{
\item{dat}{data to perform the test on. Samples on rows, features on columns}

\item{cl}{A numeric vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.}

\item{dendro}{If provided, is dendrogram providing hierarchical clustering of clusters in cl; mainly useful to speed up calculations if clusterHclust has already been called on the (cl,dat) pair. The default (NULL) is to recalculate it with the given (cl,dat) pair.}

\item{mergeMethod}{method for calculating proportion of non-null that will be used to merge clusters (if 'none', no merging will be done). See details for description of methods.}

\item{cutoff}{cutoff for merging clusters based on the proportion of non-nulls in the comparison of the clusters (i.e. value between 0,1, where lower values will make it harder to merge clusters).}

\item{plotType}{what type of plotting of dendrogram. If 'all', then all the estimates of proportion non-null will be plotted; if 'mergeMethod', then only the value used in the merging is plotted for each node.}

\item{countData}{logical as to whether input data is a count matrix (in which case log(count+1) will be used for \code{\link{clusterHclust}} and voom correction will be used in \code{\link{getBestGenes}}).}

\item{...}{arguments passed to the \code{\link{plot.phylo}} function of \code{ade4} that plots the dendrogram.}
}
\value{
Returns (invisibly) a list with elements
\itemize{
\item{\code{clustering}}{a vector of length equal to nrows(dat) giving the integer-valued cluster ids for each sample. "-1" indicates the sample was not clustered.}
\item{\code{oldClToNew}}{A table of the old cluster labels to the new cluster labels}
\item{\code{propDE}}{A table of the proportions that are DE on each node}
\item{\code{originalClusterDendro}}{The dendrogram on which the merging was based (based on the original clustering)}
}
}
\description{
Takes an input of hierarchical clusterings of clusters and returns estimates of number of proportion of non-null and merges those below a certain cutoff.
}
\details{
"JC" refers to the method of Ji and Cai (2007), and implementation of "JC" method is copied from code available on Jiashin Ji's website, December 16, 2015 (http://www.stat.cmu.edu/~jiashun/Research/software/NullandProp/). "locfdr" refers to the method of Efron (2004) and is implemented in the package \code{\link{locfdr}}. "MB" refers to the method of Meinshausen and Buhlmann (2005) and is implemented in the package \code{\link{howmany}}. "adjP" refers to the proportion of genes that are found significant based on a FDR adjusted p-values (method "BH") and a cutoff of 0.05.

If \code{mergeMethod} is not equal to 'none' then the plotting will indicate the merged clusters (assuming \code{plotType} is not 'none').
}
\examples{
data(simData)
 #create a clustering, for 8 clusters (truth was 3)
 cl<-clusterAll(simData,clusterFunction="pam",subsample=FALSE,
 sequential=FALSE, clusterDArgs=list(k=8))$cl
#merge clusters with plotting. Note argument 'use.edge.length' can improve readability
mergeResults<-mergeClusters(simData,cl=cl,plot=TRUE,plotType="all",
mergeMethod="adjP",use.edge.length=FALSE,countData=FALSE)
#compare merged to original on heatmap
hclData<-clusterHclust(dat=simData,cl,full=TRUE)
plotHeatmap(cl,heatData=simCount,clusterData=hclData,colorScale=seqPal5,
annCol=data.frame(Original=cl,Merged=mergeResults$cl,Truth=trueCluster))
}

