% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterHclust.R
\name{mergeClusters}
\alias{mergeClusters}
\title{Merge clusters based on dendrogram}
\usage{
mergeClusters(dat, cl, dendro, mergeMethod = c("none", "locfdr", "MB", "JC"),
  cutoff = 0.1)
}
\arguments{
\item{dat}{data to perform the test on. Samples on rows, features on columns}

\item{cl}{A vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.}

\item{dendro}{A dendrogram giving a hierarchical relationships between the clusters}

\item{cutoff}{A value between 0,1 at which you merge the children together (i.e. if the proportion non-null in the test is below this value, then merge the two children nodes). If set to '1', will be ignored, and only the estimated proportions will be returned}
}
\description{
Takes an input of hierarchical clusterings of clusters and returns estimates of number of proportion of non-null and merges those below a certain cutoff.
}
\details{
Implementation of "JC" method of Ji and Cai (2007) is copied from code available on Jiashin Ji's website, December 16, 2015 (http://www.stat.cmu.edu/~jiashun/Research/software/NullandProp/).
}

