% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeClusters.R
\docType{methods}
\name{mergeClusters,matrix-method}
\alias{mergeClusters}
\alias{mergeClusters,ClusterExperiment-method}
\alias{mergeClusters,matrix-method}
\title{Merge clusters based on dendrogram}
\usage{
\S4method{mergeClusters}{matrix}(x, cl, dendro = NULL,
  mergeMethod = c("none", "adjP", "locfdr", "MB", "JC"), cutoff = 0.1,
  plotType = c("none", "all", "mergeMethod"), countData = TRUE, ...)

\S4method{mergeClusters}{ClusterExperiment}(x, ...)
}
\arguments{
\item{x}{data to perform the test on. It can be a matrix or a
\code{\link{ClusterExperiment}}.}

\item{cl}{A numeric vector with cluster assignments to compare to. ``-1''
indicates the sample was not assigned to a cluster.}

\item{dendro}{dendrogram providing hierarchical clustering of clusters in cl;
The default for matrix (NULL) is to recalculate it with the given (x, cl)
pair. If x is a \code{\link{ClusterExperiment}} object, the dendrogram in
the slot \code{dendro_clusters} will be used. This means that
\code{\link{makeDendrogram}} needs to be called before
\code{mergeClusters}.}

\item{mergeMethod}{method for calculating proportion of non-null that will be
used to merge clusters (if 'none', no merging will be done). See details
for description of methods.}

\item{cutoff}{cutoff for merging clusters based on the proportion of
non-nulls in the comparison of the clusters (i.e. value between 0, 1, where
lower values will make it harder to merge clusters).}

\item{plotType}{what type of plotting of dendrogram. If 'all', then all the
estimates of proportion non-null will be plotted; if 'mergeMethod', then
only the value used in the merging is plotted for each node.}

\item{countData}{logical as to whether input data is a count matrix (in which
case \code{log(count + 1)} will be used for \code{\link{makeDendrogram}}
and voom correction will be used in \code{\link{getBestFeatures}}). Ignored
if input is \code{\link{ClusterExperiment}}.}

\item{...}{arguments passed to the \code{\link{plot.phylo}} function of
\code{ade4} that plots the dendrogram.}
}
\value{
If `x` is a matrix, it returns (invisibly) a list with elements
  \itemize{ \item{\code{clustering}}{ a vector of length equal to ncol(x)
  giving the integer-valued cluster ids for each sample. "-1" indicates the
  sample was not clustered.} \item{\code{oldClToNew}}{ A table of the old
  cluster labels to the new cluster labels.} \item{\code{propDE}}{ A table of
  the proportions that are DE on each node.}
  \item{\code{originalClusterDendro}}{ The dendrogram on which the merging
  was based (based on the original clustering).} }

If `x` is a \code{\link{ClusterExperiment}}, it returns a new
  \code{ClusterExperiment} object with an additional clustering based on the
  merging. This becomes the new primary clustering.
}
\description{
Takes an input of hierarchical clusterings of clusters and
  returns estimates of number of proportion of non-null and merges those
  below a certain cutoff.
}
\details{
"JC" refers to the method of Ji and Cai (2007), and implementation
  of "JC" method is copied from code available on Jiashin Ji's website,
  December 16, 2015
  (http://www.stat.cmu.edu/~jiashun/Research/software/NullandProp/). "locfdr"
  refers to the method of Efron (2004) and is implemented in the package
  \code{\link{locfdr}}. "MB" refers to the method of Meinshausen and Buhlmann
  (2005) and is implemented in the package \code{\link{howmany}}. "adjP"
  refers to the proportion of genes that are found significant based on a FDR
  adjusted p-values (method "BH") and a cutoff of 0.05.

If \code{mergeMethod} is not equal to 'none' then the plotting will
  indicate the merged clusters (assuming \code{plotType} is not 'none').
}
\examples{
data(simData)

#create a clustering, for 8 clusters (truth was 3)
cl<-clusterSingle(simData, clusterFunction="pam", subsample=FALSE,
sequential=FALSE, clusterDArgs=list(k=8))

#make dendrogram
cl <- makeDendrogram(cl)

#merge clusters with plotting. Note argument 'use.edge.length' can improve
#readability
merged <- mergeClusters(cl, plot=TRUE, plotType="all",
mergeMethod="adjP", use.edge.length=FALSE)

#compare merged to original
table(primaryCluster(cl), primaryCluster(merged))
}

