% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineClusters.R
\docType{methods}
\name{pipelineClusters}
\alias{pipelineClusterDetails}
\alias{pipelineClusterDetails,ClusterExperiment-method}
\alias{pipelineClusterTable}
\alias{pipelineClusterTable,ClusterExperiment-method}
\alias{pipelineClusters}
\title{Methods for pipeline clusters}
\usage{
\S4method{pipelineClusters}{ClusterExperiment}(x, iteration = 0)

\S4method{pipelineClusterDetails}{ClusterExperiment}(x)

\S4method{pipelineClusterTable}{ClusterExperiment}(x)
}
\arguments{
\item{x}{a \code{\link{ClusterExperiment}} object.}
}
\value{
\code{pipelineClusters} returns a matrix consisting of the
  appropriate columns of the \code{clusterMatrix} slot.

\code{pipelineClusterDetails} returns a \code{data.frame} with some
  details on the clusterings, such as the type (e.g., `clusterMany`,
  `combineMany`) and iteration.

\code{pipelineClusterTable} returns a table of how many of the
  clusterings belong to each of the following possible values: `final`,
  `mergeClusters`, `combineMany` and `clusterMany`.
}
\description{
The main pipeline of the package is made of
  \code{\link{clusterMany}}, \code{\link{combineMany}}, and
  \code{\link{mergeClusters}}. The clusterings from these functions (and not
  those obtained in a different way) can be obtained with the functions
  documented here.
}
\examples{
data(simData)

cl <- clusterMany(simData,nPCADims=c(5,10,50),  dimReduce="PCA",
clusterFunction="pam", ks=2:4, findBestK=c(FALSE), removeSil=TRUE,
subsample=FALSE)

clCommon <- combineMany(cl, whichClusters="pipeline", proportion=0.7,
minSize=10)

clCommon <- makeDendrogram(clCommon)

clMerged <- mergeClusters(clCommon)

head(pipelineClusters(clMerged))
pipelineClusterDetails(clMerged)
pipelineClusterTable(clMerged)
}

