% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarplot.R
\docType{methods}
\name{plotBarplot,ClusterExperiment,character-method}
\alias{plotBarplot,ClusterExperiment,character-method}
\alias{plotBarplot}
\alias{plotBarplot,ClusterExperiment,missing-method}
\alias{plotBarplot,ClusterExperiment,numeric-method}
\alias{plotBarplot,ClusterExperiment,missing-method}
\alias{plotBarplot,vector,missing-method}
\alias{plotBarplot,matrix,missing-method}
\title{Barplot of 1 or 2 clusterings}
\usage{
\S4method{plotBarplot}{ClusterExperiment,character}(clusters, whichClusters,
  ...)

\S4method{plotBarplot}{ClusterExperiment,missing}(clusters, whichClusters, ...)

\S4method{plotBarplot}{ClusterExperiment,numeric}(clusters, whichClusters,
  labels = c("names", "ids"), ...)

\S4method{plotBarplot}{ClusterExperiment,missing}(clusters, whichClusters, ...)

\S4method{plotBarplot}{vector,missing}(clusters, whichClusters, ...)

\S4method{plotBarplot}{matrix,missing}(clusters, whichClusters, xNames = NULL,
  legNames = NULL, legend = TRUE, xlab = NULL, legend.title = NULL,
  unassignedColor = "white", missingColor = "grey",
  colPalette = bigPalette, ...)
}
\arguments{
\item{clusters}{A matrix of with each column corresponding to a clustering
and each row a sample or a \code{\link{ClusterExperiment}} object.}

\item{...}{for \code{plotBarplot} arguments passed either to the method
of \code{plotBarplot} for matrices or ultimately to \code{\link{barplot}}.}

\item{labels}{if clusters is a clusterExperiment object, then labels defines
whether the clusters will be identified by their names values in
clusterLegend (labels="names", the default) or by their clusterIds value in
clusterLegend (labels="ids").}

\item{xNames}{names for the first clusters (on x-axis). By default uses
values in 1st cluster of clusters matrix}

\item{legNames}{names for the first clusters (in legend). By default uses
values in 2nd cluster of clusters matrix}

\item{legend}{whether to plot the legend}

\item{xlab}{label for x-axis. By default or if equal NULL the column name of
the 1st cluster of clusters matrix}

\item{legend.title}{label for legend. By default or if equal NULL the column
name of the 2st cluster of clusters matrix}

\item{colPalette}{a vector of colors used for the different clusters. Must be
as long as the maximum number of clusters found in any single
clustering/column given in \code{clusters} or will otherwise return an
error.}
}
\description{
Make a barplot of sample's assignments to clusters for single clustering, or
cross comparison for two clusterings.
}
\details{
The first column of the cluster matrix will be on the x-axis and the
  second column will separate the groups of the first column.

All arguments of the matrix version can be passed to the
  \code{ClusterExperiment} version. As noted above, however, some arguments
  have different interpretations.

If \code{whichClusters = "workflow"}, then the most recent two
  clusters of the workflow will be chosen where recent is based on the
  following order (most recent first): final, mergeClusters, combineMany,
  clusterMany.
}
\examples{
#clustering using pam: try using different dimensions of pca and different k
data(simData)

cl <- clusterMany(simData, nPCADims=c(5, 10, 50), dimReduce="PCA",
clusterFunction="pam", ks=2:4, findBestK=c(TRUE,FALSE),
removeSil=c(TRUE,FALSE))

plotBarplot(cl)
plotBarplot(cl,whichClusters=1:2)

}
\author{
Elizabeth Purdom
}
