% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustersWorkflow.R
\docType{methods}
\name{plotClustersWorkflow,ClusterExperiment-method}
\alias{plotClustersWorkflow,ClusterExperiment-method}
\alias{plotClustersWorkflow}
\title{A plot of clusterings specific for clusterMany and workflow
  visualization}
\usage{
\S4method{plotClustersWorkflow}{ClusterExperiment}(object,
  whichClusters = c("mergeClusters", "combineMany"),
  whichClusterMany = NULL, nBlankLines = ceiling(nClusters(object) * 0.05),
  nSizeResult = ceiling(nClusters(object) * 0.02), clusterLabels = TRUE,
  clusterManyLabels = TRUE, sortBy = c("highlighted", "clusterMany"),
  resultsOnTop = TRUE, ...)
}
\arguments{
\item{object}{A \code{ClusterExperiment} object on which
\code{\link{clusterMany}} has been run}

\item{whichClusters}{which clusterings to "highlight", i.e draw separately 
from the \code{clusterMany} results. Can be numeric or character vector, 
indicating the indices or clusterLabels/clusterTypes of the clusterings of 
interest, respectively.}

\item{whichClusterMany}{numeric indices of which of the clusterMany 
clusterings to plot (if NULL, defaults to all). Unlike
\code{whichClusters}, these must be numeric indices. They must also refer
to clusterings of clusterType \code{clusterMany}.}

\item{nBlankLines}{the number of blank (i.e. white) rows to add between the
clusterMany clusterings and the highlighted clusterings.}

\item{nSizeResult}{the number of rows each highlighted clustering should take up.
Increasing the number increases the thickness of the rectangles
representing the highlighted clusterings.}

\item{clusterLabels}{either logical, indicating whether to plot the labels
for the clusterings identified to be highlighted in the
\code{whichClusters} argument, or a character vector of labels to use.}

\item{clusterManyLabels}{either logical, indicating whether to plot the
labels for the clusterings from clusterMany identified in the
\code{whichClusterMany}, or a character vector of labels to use.}

\item{sortBy}{how to align the clusters. If "highlighted" then the
highlighted clusters indicated in the argument \code{whichClusters} are
first in the alignment done by \code{plotClusters}. If "clusterMany", then
the clusterMany results are first in the alignment. (Note this does not
determine where they will be plotted, but how they are ordered in the
aligning step done by \code{plotClusters})}

\item{resultsOnTop}{logical. Whether the results should be plotted on the top
of clusterMany results or the bottom.
@param ... arguments passed to \code{\link{plotClusters}}
@details This plot is solely intended to make it easier to use the 
  \code{\link{plotClusters}} visualization when there are a large number of
  clusterings from a call to \code{\link{clusterMany}}. This plot separates
  out the \code{\link{clusterMany}} results from a designated clustering of
  interest, as indicated by the \code{whichClusters} argument
  (by default clusterings from a call to \code{\link{combineMany}} or 
  \code{\link{mergeClusters}}). In addition the highlighted clusters are
  made bigger so that they can be easily seen.}
}
\description{
A realization of \code{\link{plotClusters}} call specific to
  separating out the results of \code{clusterMany} and other clustering
  results.
}
\seealso{
\code{\link{plotClusters}}, \code{\link{clusterMany}}
}
