% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTableClusters.R
\docType{methods}
\name{plotTableClusters}
\alias{plotTableClusters}
\alias{plotTableClusters,ClusterExperiment-method}
\alias{plotTableClusters,table-method}
\alias{tableClusters,ClusterExperiment,character-method}
\alias{tableClusters}
\alias{tableClusters,ClusterExperiment,missing-method}
\alias{tableClusters,ClusterExperiment,numeric-method}
\title{Plot heatmap of cross-tabs of 2 clusterings}
\usage{
\S4method{plotTableClusters}{ClusterExperiment}(object, whichClusters,
  ignoreUnassigned = FALSE, margin = NA, ...)

\S4method{plotTableClusters}{table}(object, clusterLegend = NULL,
  cluster = FALSE, ...)

\S4method{tableClusters}{ClusterExperiment,character}(object, whichClusters,
  ...)

\S4method{tableClusters}{ClusterExperiment,missing}(object, whichClusters, ...)

\S4method{tableClusters}{ClusterExperiment,numeric}(object, whichClusters,
  useNames = TRUE, ...)
}
\arguments{
\item{object}{ClusterExperiment object (or matrix with table result)}

\item{whichClusters}{which clusters to tabulate. For \code{plotTableClusters} should
be 2 clusters, for \code{tableClusters} can indicate arbitrary number.}

\item{ignoreUnassigned}{logical as to whether to ignore unassigned clusters in the 
plotting. This means they will also be ignored in the calculations of the proportions 
(if \code{margin} not NA).}

\item{margin}{if NA, the counts from \code{tableClusters} will be plotted. Otherwise, 
\code{\link[base]{prop.table}} will be called and the argument \code{margin} #' will be passed to \code{prop.table} to determine how proportions should be calculated.}

\item{...}{arguments passed on to \code{plotHeatmap}}

\item{clusterLegend}{list in \code{clusterLegend} format that gives colors for the
clusters tabulated.}

\item{cluster}{logical, whether to cluster the rows and columns of the table. Passed
to arguments \code{clusterFeatures} AND \code{clusterSamples} of \code{plotHeatmap}.}

\item{useNames}{for \code{tableClusters}, whether the output should be tabled
with names (\code{useNames=TRUE}) or ids (\code{useNames=FALSE})}
}
\description{
Plot heatmap of cross-tabulations of two clusterings
}
\details{
Note that the cluster labels in \code{plotTableClusters} and 
\code{tableClusters} are converted to "proper" R names via \code{make.names}. This is 
because \code{tableClusters} calls the R function \code{table}, which makes this 
conversion
}
\examples{
#clustering using pam: try using different dimensions of pca and different k
data(simData)

cl <- clusterMany(simData, nReducedDims=c(5, 10, 50), reducedDim="PCA",
clusterFunction="pam", ks=2:4, findBestK=c(TRUE,FALSE),
removeSil=c(TRUE,FALSE))
#give arbitrary names to clusters for demonstration
cl<-renameClusters(cl,value=letters[1:nClusters(cl)[1]],whichCluster=1)
tableClusters(cl,whichClusters=1:2)
#heatmap of the counts in each entry of table:
plotTableClusters(cl,whichClusters=1:2, ignoreUnassigned=TRUE)
}
\seealso{
\code{\link[base]{prop.table}}

\code{\link{plotHeatmap}}

\code{\link[base]{table}}
}
