% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFilterStats.R, R/makeReducedDims.R
\docType{methods}
\name{reduceFunctions}
\alias{reduceFunctions}
\alias{makeFilterStats}
\alias{makeFilterStats,SummarizedExperiment-method}
\alias{makeFilterStats,matrixOrHDF5-method}
\alias{makeFilterStats,ClusterExperiment-method}
\alias{listBuiltInFilterStats}
\alias{filterData,SummarizedExperiment-method}
\alias{filterData}
\alias{defaultNDims,SingleCellExperiment-method}
\alias{defaultNDims}
\alias{filterNames,SummarizedExperiment-method}
\alias{filterNames}
\alias{makeReducedDims,SingleCellExperiment-method}
\alias{makeReducedDims}
\alias{makeReducedDims,matrixOrHDF5-method}
\alias{makeReducedDims,SummarizedExperiment-method}
\alias{makeReducedDims,ClusterExperiment-method}
\alias{listBuiltInReducedDims}
\title{Filtering statistics and Dimensionality Reduction Functions}
\usage{
\S4method{makeFilterStats}{SummarizedExperiment}(object,
  filterStats = listBuiltInFilterStats(), transFun = NULL,
  isCount = FALSE)

\S4method{makeFilterStats}{matrixOrHDF5}(object, ...)

\S4method{makeFilterStats}{ClusterExperiment}(object,
  whichClusterIgnoreUnassigned = NULL,
  filterStats = listBuiltInFilterStats(), ...)

listBuiltInFilterStats()

\S4method{filterData}{SummarizedExperiment}(object, filterStats, cutoff,
  percentile, absolute = FALSE, keepLarge = TRUE)

\S4method{defaultNDims}{SingleCellExperiment}(object, reduceMethod, typeToShow)

\S4method{filterNames}{SummarizedExperiment}(object)

\S4method{makeReducedDims}{SingleCellExperiment}(object, reducedDims = "PCA",
  maxDims = 500, transFun = NULL, isCount = FALSE)

\S4method{makeReducedDims}{matrixOrHDF5}(object, ...)

\S4method{makeReducedDims}{SummarizedExperiment}(object, ...)

\S4method{makeReducedDims}{ClusterExperiment}(object, ...)

listBuiltInReducedDims()
}
\arguments{
\item{object}{object from which user wants to calculate per-row statistics}

\item{filterStats}{character vector of statistics to calculate. 
Must be one of the character values given by \code{listBuildInFilterStats()}.}

\item{transFun}{a transformation function to be applied to the data. If the transformation 
applied to the data creates an error or NA values, then the function will throw an error.
If object is of class \code{ClusterExperiment}, the stored transformation will be used 
and giving this parameter will result in an error.}

\item{isCount}{if \code{transFun=NULL}, then \code{isCount=TRUE} will determine the
will give a transformation function \code{function(x){x}}. Ignored if \code{transFun=NULL}.
If object is of class \code{ClusterExperiment}, the stored transformation will be used 
and giving this parameter will result in an error.}

\item{...}{Values passed on the the 'SingleCellExperiment' method.}

\item{whichClusterIgnoreUnassigned}{indicates clustering that should be used to filter out unassigned samples from the calculations. If \code{NULL} no filtering of samples will be done. See details for more information.}

\item{cutoff}{numeric. A value at which to filter the rows (genes) for the test statistic}

\item{percentile}{numeric. Either a number between 0,1 indicating what percentage of the rows (genes) to keep or an integer value indicated the number of rows (genes) to keep}

\item{absolute}{whether to take the absolute value of the filter statistic}

\item{keepLarge}{logical whether to keep rows (genes) with large values of the test statistic or small values of the test statistic.}

\item{reduceMethod}{character. A method or methods for reducing the size of the data, either by filtering the rows (genes) or by a dimensionality reduction method. Must either be 1) must match the name of a built-in method, in which case if it is not already existing in the object will be passed to \code{\link{makeFilterStats}} or \code{link{makeReducedDims}}, or 2) must match a stored filtering statistic or dimensionality reduction in the object}

\item{typeToShow}{character (optional). If given, should be one of "filterStats" or "reducedDims" to indicate of the values in the reduceMethod vector, only show those corresponding to "filterStats" or "reducedDims" options.}

\item{reducedDims}{a vector of character values indicating the methods of dimensionality reduction to be performed. Currently only "PCA" is implemented.}

\item{maxDims}{Numeric vector of integer giving the number of PC dimensions to calculate. 
\code{maxDims} can also take values between (0,1) to indicate keeping the
number of dimensions necessary to account for that proportion of the
variance. \code{maxDims} should be of same length as \code{reducedDims}, indicating 
the number of dimensions to keep for each method (if \code{maxDims} is of length 1, 
the same number of dimensions will be used for each).}
}
\value{
\code{makeFilterStats} returns a \code{\link[SummarizedExperiment]{SummarizedExperiment}} object with the requested filtering statistics will be added to the \code{DataFrame} in the \code{rowData} slot and given names corresponding to the \code{filterStats} values. Warning: the function will overwrite existing columns in \code{rowData} with the same name. Columns in the \code{rowData} slot with different names should not be affected.

\code{filterData} returns a SingleCellExperiment object with the rows (genes) removed based on filters

\code{defaultNDims} returns a numeric vector giving the default dimensions the methods in \code{clusterExperiment} will use for reducing the size of the data. If \code{typeToShow} is missing, the resulting vector will be equal to the length of \code{reduceMethod}. Otherwise, it will be a vector with all the unique valid default values for the \code{typeToShow} (note that different dimensionality reduction methods can have different maximal dimensions, so the result may not be of length one in this case).

\code{filterNames} returns a vector of the columns of the rowData that are considered valid filtering statistics. Currently any numeric column in rowData is a valid filtering statistic.

\code{makeReducedDims} returns a \code{\link{SingleCellExperiment}} containing the calculated dimensionality reduction in the \code{reduceDims} with names corresponding to the name given in \code{reducedDims}.
}
\description{
Functions for calculating and manipulating either filtering statistics, stored in rowData, or the dimensionality reduction results, stored in reducedDims.
}
\details{
\code{whichClusterIgnoreUnassigned} is only an option when applied to a
\code{ClusterExperiment} classs and indicates that the filtering statistics should be
calculated based on samples that are unassigned by the
 designated clustering. The name given to the filter in this case is of the form
 \code{<filterStats>_<clusterLabel>}, i.e. the clustering label of the clustering is
 appended to the standard name for the filtering statistic.

Note that \code{filterData} returns a SingleCellExperiment object. To get the actual data out use either assay or \code{\link{transformData}} if transformed data is desired.

The PCA method uses either \code{prcomp} from the \code{stats} package or  \code{svds} from the \code{RSpectra} package to perform PCA. Both are called on \code{t(assay(x))} with \code{center=TRUE} and \code{scale=TRUE} (i.e. the feature are centered and scaled), so that
  it is performing PCA on the correlation matrix of the features.
}
\examples{
data(simData)
listBuiltInFilterStats()
scf<-makeFilterStats(simData,filterStats=c("var","mad"))
scf
scfFiltered<-filterData(scf,filterStats="mad",percentile=10)
scfFiltered
assay(scfFiltered)[1:10,1:10]
data(simData)
listBuiltInReducedDims()
scf<-makeReducedDims(simData, reducedDims="PCA", maxDims=3)
scf
}
