% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelper.R
\docType{methods}
\name{renameClusters,ClusterExperiment,character-method}
\alias{renameClusters,ClusterExperiment,character-method}
\alias{renameClusters}
\alias{recolorClusters,ClusterExperiment,character-method}
\alias{recolorClusters}
\title{Change assigned names or colors of clusters}
\usage{
\S4method{renameClusters}{ClusterExperiment,character}(object, value,
  whichCluster = "primary", matchTo = c("name", "clusterIds"))

\S4method{recolorClusters}{ClusterExperiment,character}(object, value,
  whichCluster = "primary", matchTo = c("name", "clusterIds"))
}
\arguments{
\item{whichCluster}{argument to identify cluster, taking input like 
\code{whichClusters}, only requires that only a single clustering can be identified.}
}
\value{
\code{renameClusters} changes the names assigned to clusters within a clustering

\code{recolorClusters} changes the colors assigned to clusters within a clustering
}
\description{
Change the assigned names or colors of the clusters in a clustering
  stored in the clusterLegend slot of the object.
}
