% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsec.R
\docType{methods}
\name{RSEC}
\alias{RSEC}
\alias{RSEC-methods}
\alias{RSEC,ClusterExperiment-method}
\alias{RSEC,matrix-method}
\alias{RSEC,SummarizedExperiment-method}
\alias{RSEC,data.frame-method}
\alias{RSEC,ClusterExperiment-method}
\title{Resampling-based Sequential Ensemble Clustering}
\usage{
\S4method{RSEC}{matrix}(x, isCount = FALSE, transFun = NULL,
  dimReduce = "PCA", nVarDims = NA, nPCADims = c(50), k0s = 4:15,
  clusterFunction = "hierarchical01", alphas = c(0.1, 0.2, 0.3),
  betas = 0.9, minSizes = 1, combineProportion = 0.7,
  combineMinSize = 5, dendroReduce = "mad", dendroNDims = 1000,
  mergeMethod = "adjP", mergeCutoff = 0.05, verbose = FALSE,
  mainClusterArgs = NULL, subsampleArgs = NULL, seqArgs = NULL,
  ncores = 1, random.seed = NULL, run = TRUE)

\S4method{RSEC}{SummarizedExperiment}(x, ...)

\S4method{RSEC}{data.frame}(x, ...)

\S4method{RSEC}{ClusterExperiment}(x, eraseOld = FALSE,
  rerunClusterMany = FALSE, ...)
}
\arguments{
\item{x}{the data matrix on which to run the clustering. Can be: matrix (with
genes in rows), a list of datasets overwhich the clusterings should be run,
a \code{SummarizedExperiment} object, or a \code{ClusterExperiment} object.}

\item{isCount}{logical. Whether the data are in counts, in which case the
default \code{transFun} argument is set as log2(x+1). This is simply a
convenience to the user, and can be overridden by giving an explicit
function to \code{transFun}.}

\item{transFun}{function A function to use to transform the input data matrix
before clustering.}

\item{dimReduce}{character A character identifying what type of
dimensionality reduction to perform before clustering. Options are
"none","PCA", "var","cv", and "mad". See \code{\link{transform}} for more
details.}

\item{nVarDims}{vector of the number of the most variable features to keep
(when "var", "cv", or "mad" is identified in \code{dimReduce}). If NA is
included, then the full dataset will also be included.}

\item{nPCADims}{vector of the number of PCs to use (when 'PCA' is identified
in \code{dimReduce}). If NA is included, then the full dataset will also be
included.}

\item{k0s}{the k0 parameter for sequential clustering (see
\code{\link{seqCluster}})}

\item{clusterFunction}{function used for the clustering. Note that unlike in
\code{\link{clusterSingle}}, this must be a character vector of pre-defined
clustering techniques, and can not be a user-defined function. Current
functions can be found by typing \code{listBuiltInFunctions()} into the
command-line.}

\item{alphas}{values of alpha to be tried. Only used for clusterFunctions of
type '01'. Determines tightness required in creating clusters from the
dissimilarity matrix. Takes on values in [0,1]. See documentation of
\code{\link{ClusterFunction}}.}

\item{betas}{values of \code{beta} to be tried in sequential steps. Only used
for \code{sequential=TRUE}. Determines the similarity between two clusters
required in order to deem the cluster stable. Takes on values in [0,1]. See
documentation of \code{\link{seqCluster}}.}

\item{minSizes}{the minimimum size required for a cluster (in the
\code{mainClustering} step). Clusters smaller than this are not kept and samples
are left unassigned.}

\item{combineProportion}{passed to \code{proportion} in
\code{\link{combineMany}}}

\item{combineMinSize}{passed to \code{minSize} in \code{\link{combineMany}}}

\item{dendroReduce}{passed to \code{dimReduce} in
\code{\link{makeDendrogram}}}

\item{dendroNDims}{passed to \code{ndims} in \code{\link{makeDendrogram}}}

\item{mergeMethod}{passed to \code{mergeMethod} in
\code{\link{mergeClusters}}}

\item{mergeCutoff}{passed to \code{cutoff} in \code{\link{mergeClusters}}}

\item{verbose}{logical. If TRUE it will print informative messages.}

\item{mainClusterArgs}{list of arguments to be passed for the mainClustering step, see
help pages of \code{\link{mainClustering}}.}

\item{subsampleArgs}{list of arguments to be passed to the subsampling step
(if \code{subsample=TRUE}), see help pages of
\code{\link{subsampleClustering}}.}

\item{seqArgs}{list of arguments to be passed to \code{\link{seqCluster}}.}

\item{ncores}{the number of threads}

\item{random.seed}{a value to set seed before each run of clusterSingle (so
that all of the runs are run on the same subsample of the data). Note, if
'random.seed' is set, argument 'ncores' should NOT be passed via
subsampleArgs; instead set the argument 'ncores' of clusterMany directly
(which is preferred for improving speed anyway).}

\item{run}{logical. If FALSE, doesn't run clustering, but just returns matrix
of parameters that will be run, for the purpose of inspection by user (with
rownames equal to the names of the resulting column names of clMat object
that would be returned if \code{run=TRUE}). Even if \code{run=FALSE},
however, the function will create the dimensionality reductions of the data
indicated by the user input.}

\item{...}{For signature \code{list}, arguments to be passed on to mclapply
(if ncores>1). For all the other signatures, arguments to be passed to the
method for signature \code{list}.}

\item{eraseOld}{logical. Only relevant if input \code{x} is of class
\code{ClusterExperiment}. If TRUE, will erase existing workflow results
(clusterMany as well as mergeClusters and combineMany). If FALSE, existing
workflow results will have "\code{_i}" added to the clusterTypes value,
where \code{i} is one more than the largest such existing workflow
clusterTypes.}

\item{rerunClusterMany}{logical. If the object is a clusterExperiment object,
determines whether to rerun the clusterMany step. Useful if want to try
different parameters for combining clusters after the clusterMany step,
without the computational costs of the clusterMany step.}
}
\description{
Implementation of the RSEC algorithm (Resampling-based
  Sequential Ensemble Clustering) for single cell sequencing data. This is a
  wrapper function around the existing clusterExperiment workflow that
  results in the output of RSEC.
}
