% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCreation.R
\docType{data}
\name{rsecFluidigm}
\alias{rsecFluidigm}
\title{RSEC run for vignette}
\format{ClusterExperiment object, the result of running \code{\link{RSEC}} on
fluidigm data described in vignette and available in the \code{scRNAseq}
package.}
\description{
RSEC run for vignette
}
\examples{
#code used to create rsecFluidigm:
\dontrun{
library(scRNAseq)
data("fluidigm")
se <- fluidigm[,colData(fluidigm)[,"Coverage_Type"]=="High"]
wh_zero <- which(rowSums(assay(se))==0)
pass_filter <- apply(assay(se), 1, function(x) length(x[x >= 10]) >= 10)
se <- se[pass_filter,]
fq <- round(limma::normalizeQuantiles(assay(se)))
assays(se) <- list(normalized_counts=fq)
wh<-which(colnames(colData(se)) \%in\% c("Cluster1","Cluster2"))
colnames(colData(se))[wh]<-c("Published1","Published2")
library(clusterExperiment)
ncores<-1
system.time(
  rsecFluidigm<-RSEC(se, 
                     isCount = TRUE, 
                     k0s = 4:15, 
                     alphas=c(0.1, 0.2, 0.3), 
                     betas = 0.9,
                     reduceMethod="PCA", 
                     nReducedDims=10,
                     minSizes=1, 
                     clusterFunction="hierarchical01",
                     consensusMinSize=3,
                     consensusProportion=0.7,
                     dendroReduce= "mad",
                     dendroNDims=1000,
                     mergeMethod="adjP",
                     mergeCutoff=0.01,
                     ncores=1, 
                     random.seed=176201)
)
metadata(rsecFluidigm)$packageVersion<-packageVersion("clusterExperiment")
if(length(tableClusters(rsecFluidigm,whichCluster="mergeClusters")) != 6) 
 stop("rsecFluidigm object has changed -- mergeClusters")
if(length(tableClusters(rsecFluidigm,whichCluster="makeConsensus")) != 8)
 stop("rsecFluidigm object has changed -- makeConsensus")
packageVersion("clusterExperiment")
devtools::use_data(rsecFluidigm,overwrite=FALSE)
}
}
\seealso{
\code{\link[scRNAseq]{fluidigm}}
}
\author{
Elizabeth Purdom \email{epurdom@stat.berkeley.edu}
}
\keyword{data}
