% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCreation.R
\docType{data}
\name{rsecFluidigm}
\alias{rsecFluidigm}
\title{RSEC run for vignette}
\format{ClusterExperiment object, the result of running \code{\link{RSEC}} on 
fluidigm data described in vignette and available in the \code{scRNAseq} 
package.}
\description{
RSEC run for vignette
}
\examples{
#code used to create rsecFluidigm:
\dontrun{
library(scRNAseq)
data("fluidigm")
se <- fluidigm[,colData(fluidigm)[,"Coverage_Type"]=="High"]
wh_zero <- which(rowSums(assay(se))==0)
pass_filter <- apply(assay(se), 1, function(x) length(x[x >= 10]) >= 10)
se <- se[pass_filter,]
fq <- round(limma::normalizeQuantiles(assay(se)))
assays(se) <- list(normalized_counts=fq)
wh<-which(colnames(colData(se)) \%in\% c("Cluster1","Cluster2"))
colnames(colData(se))[wh]<-c("Published1","Published2")
library(ClusterExperiment)
ncores<-1
system.time(rsecFluidigm<-RSEC(se, isCount = TRUE,dimReduce="PCA",nDimReduce=10,
ncores=ncores,random.seed=176201, clusterFunction="hierarchical01",
 combineMinSize=3))
packageVersion("clusterExperiment")
devtools::use_data(rsecFluidigm,overwrite=FALSE)
}
}
\seealso{
\code{\link[scRNAseq]{fluidigm}}
}
\author{
Elizabeth Purdom \email{epurdom@stat.berkeley.edu}
}
\keyword{data}
