% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorDefinitions.R
\docType{data}
\name{showHeatmapPalettes}
\alias{seqPal1}
\alias{seqPal2}
\alias{seqPal3}
\alias{seqPal4}
\alias{seqPal5}
\alias{showHeatmapPalettes}
\title{Set of colors useful for heatmap gradients}
\format{\preformatted{ chr [1:16] "#000000" "#00003B" "#000077" "#00009E" "#0000C2" "#0826CD" "#135CCD" "#247AB5" "#39848B" ...
}}
\usage{
showHeatmapPalettes()

seqPal5

seqPal2

seqPal3

seqPal4

seqPal1
}
\description{
Set of colors useful for heatmap gradients
}
\details{
seqPal1-seqPal4 are palettes for the heatmap. showHeatmapPalettes() will show you
these palettes.
}
\examples{
#show the palette colors
showPalettes()
#compare the palettes on heatmap
load(simData)
load(simCount)
par(mfrow=c(2,3))
dualHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal1,main="seqPal1")
dualHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal2,main="seqPal2")
dualHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal3,main="seqPal3")
dualHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal4,main="seqPal4")
dualHeatmap(cl,heatData=simCount,clusterData=simData,colorScale=seqPal5,main="seqPal5")
}
\keyword{datasets}

