% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorDefinitions.R
\docType{data}
\name{showHeatmapPalettes}
\alias{seqPal1}
\alias{seqPal2}
\alias{seqPal3}
\alias{seqPal4}
\alias{seqPal5}
\alias{showHeatmapPalettes}
\title{Set of colors useful for heatmap gradients}
\format{\preformatted{ chr [1:16] "#000000" "#00003B" "#000077" "#00009E" "#0000C2" ...
}}
\usage{
showHeatmapPalettes()

seqPal5

seqPal2

seqPal3

seqPal4

seqPal1
}
\description{
Set of colors useful for heatmap gradients
}
\details{
seqPal1-seqPal4 are palettes for the heatmap. showHeatmapPalettes() will show you
these palettes.
}
\examples{
#show the palette colors
showPalettes()
}
\keyword{datasets}

