% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFilter.R
\docType{methods}
\name{SingleCellFilter-methods}
\alias{SingleCellFilter-methods}
\alias{SingleCellFilter}
\alias{filterStats,SingleCellFilter,missing-method}
\alias{filterStats<-,SingleCellFilter-method}
\alias{filterNames,SingleCellFilter-method}
\alias{filterNames<-,SingleCellFilter-method}
\title{Helper methods for the SingleCellFilter class}
\usage{
\S4method{filterStats}{SingleCellFilter,character}(object, type)

\S4method{filterStats}{SingleCellFilter,missing}(object, type)

\S4method{filterStats}{SingleCellFilter}(object, type, ...) <- value

\S4method{filterNames}{SingleCellFilter}(object)

\S4method{filterNames}{SingleCellFilter}(object) <- value
}
\description{
This is a collection of helper methods for the SingleCellFilter class.
}
\details{
Note that the replacement functions never actually completely replace the slot \code{filterStats}; they update existing filters of the same name and add filters with new names to the existing filters.
}
