% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFilter.R
\docType{methods}
\name{SingleCellFilter-methods}
\alias{SingleCellFilter-methods}
\alias{filterStats,SingleCellFilter,missing-method}
\alias{filterStats<-,SingleCellFilter-method}
\alias{filterNames,SingleCellFilter-method}
\alias{filterNames<-,SingleCellFilter-method}
\alias{[,SingleCellFilter,ANY,ANY,ANY-method}
\alias{show,SingleCellFilter-method}
\title{Helper methods for the SingleCellFilter class}
\usage{
\S4method{filterStats}{SingleCellFilter,character}(object, type)

\S4method{filterStats}{SingleCellFilter,missing}(object, type)

\S4method{filterStats}{SingleCellFilter}(object, type, ...) <- value

\S4method{filterNames}{SingleCellFilter}(object)

\S4method{filterNames}{SingleCellFilter}(object) <- value

\S4method{[}{SingleCellFilter,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{SingleCellFilter}(object)
}
\arguments{
\item{..., i, j, drop}{Forwarded to the
\code{\link{SingleCellExperiment}} method.}

\item{value}{The value to be substituted in the corresponding slot. See the
slot descriptions in \code{\link{SingleCellFilter}} for details on what
objects may be passed to these functions.}
}
\description{
This is a collection of helper methods for the SingleCellFilter class.
}
\details{
Note that the replacement functions never actually completely
  replace the slot \code{filterStats} unless the replacement value is NULL
  They update existing filters of the
  same name and add filters with new names to the existing filters.
}
