% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFunction.R
\docType{methods}
\name{transformData}
\alias{transformData}
\alias{transformData,ClusterExperiment-method}
\alias{transformData,SingleCellExperiment-method}
\alias{transformData,SummarizedExperiment-method}
\title{Transform the original data in a ClusterExperiment object

Provides the transformed data (as defined by the object)}
\usage{
\S4method{transformData}{matrix}(object, transFun = NULL, isCount = FALSE)

\S4method{transformData}{ClusterExperiment}(object)

\S4method{transformData}{SingleCellExperiment}(object, ...)

\S4method{transformData}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{_data}{a matrix, SummarizedExperiment, SingleCellExperiment or ClusterExperiment object.}
}
\details{
The data matrix defined by \code{assay(x)} is transformed based on
  the transformation function defined in x.
}
\examples{
mat <- matrix(data=rnorm(200), ncol=10)
mat[1,1] <- -1 #force a negative value
labels <- gl(5, 2)
cc <- ClusterExperiment(mat, as.numeric(labels), transformation =
function(x){x^2}) #define transformation as x^2
z<-transformData(cc) 
}
