% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateObject.R
\name{updateObject}
\alias{updateObject}
\title{Update old object to current class

This function updates ClusterExperiment objects from previous versions of package}
\usage{
updateObject(object, checkTransformAndAssay = FALSE)
}
\arguments{
\item{object}{a \code{ClusterExperiment} (or \code{clusterExperiment} from older 
versions). Must have at a minimum a slot \code{clusterMatrix}.}

\item{checkTransformAndAssay}{logical. Whether to check the content of the
assay and given transformation function for whether they are valid.}
}
\value{
A valid \code{ClusterExperiment} object based on the current definition of 
ClusterExperiment.
}
