% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowClusters.R
\docType{methods}
\name{workflowClusters}
\alias{workflowClusterDetails}
\alias{workflowClusterDetails,ClusterExperiment-method}
\alias{workflowClusterTable}
\alias{workflowClusterTable,ClusterExperiment-method}
\alias{workflowClusters}
\alias{workflowClusters,ClusterExperiment-method}
\title{Methods for workflow clusters}
\usage{
\S4method{workflowClusters}{ClusterExperiment}(x, iteration = 0)

\S4method{workflowClusterDetails}{ClusterExperiment}(x)

\S4method{workflowClusterTable}{ClusterExperiment}(x)
}
\arguments{
\item{x}{a \code{\link{ClusterExperiment}} object.}

\item{iteration}{numeric. Which iteration of the workflow should be used.}
}
\value{
\code{workflowClusters} returns a matrix consisting of the
  appropriate columns of the \code{clusterMatrix} slot.

\code{workflowClusterDetails} returns a \code{data.frame} with some
  details on the clusterings, such as the type (e.g., `clusterMany`,
  `combineMany`) and iteration.

\code{workflowClusterTable} returns a table of how many of the
  clusterings belong to each of the following possible values: `final`,
  `mergeClusters`, `combineMany` and `clusterMany`.
}
\description{
The main workflow of the package is made of
  \code{\link{clusterMany}}, \code{\link{combineMany}}, and
  \code{\link{mergeClusters}}. The clusterings from these functions (and not
  those obtained in a different way) can be obtained with the functions
  documented here.
}
\examples{
data(simData)

cl <- clusterMany(simData,nPCADims=c(5,10,50),  dimReduce="PCA",
clusterFunction="pam", ks=2:4, findBestK=c(FALSE), removeSil=TRUE,
subsample=FALSE)

clCommon <- combineMany(cl, whichClusters="workflow", proportion=0.7,
minSize=10)

clCommon <- makeDendrogram(clCommon)

clMerged <- mergeClusters(clCommon)

head(workflowClusters(clMerged))
workflowClusterDetails(clMerged)
workflowClusterTable(clMerged)
}

